Refactor core/vector.rs into modular structure
Moved the 5300+ line `src/core/vector.rs` into a directory-based module `src/core/vector/` with separate files for types, sparse vectors, SIMD internals, operations, validation, metrics, and tests.

This improves maintainability and separation of concerns while preserving the public API.
