+++ src/core/vector/metric.rs
@@ -1,0 +1,239 @@
+//! Distance metric definition.
+
+use crate::utils::error::Result;
+use std::fmt;
+use super::ops::{cosine_similarity, euclidean_distance, dot_product};
+
+/// Specifies which distance or similarity metric to use for vector operations.
+///
+/// This enum provides a unified interface for computing distances and similarities
+/// between vectors, dispatching to the appropriate underlying function based on the
+/// selected metric.
+///
+/// # Choosing a Metric
+///
+/// | Metric | Best For | Range | Notes |
+/// |--------|----------|-------|-------|
+/// | [`Cosine`](DistanceMetric::Cosine) | Semantic similarity, text embeddings | [-1, 1] similarity, [0, 2] distance | Scale-invariant, most common for embeddings |
+/// | [`Euclidean`](DistanceMetric::Euclidean) | Spatial data, image features | [0, ∞) distance | Sensitive to vector magnitude |
+/// | [`DotProduct`](DistanceMetric::DotProduct) | Pre-normalized vectors, MaxIP search | (-∞, ∞) | Fastest; requires normalized vectors for cosine-like behavior |
+///
+/// # Example
+///
+/// ```rust
+/// use gallifreydb::core::vector::DistanceMetric;
+///
+/// let a = vec![1.0, 0.0, 0.0];
+/// let b = vec![0.0, 1.0, 0.0];
+///
+/// // Using cosine similarity (orthogonal vectors = 0 similarity)
+/// let similarity = DistanceMetric::Cosine.compute_similarity(&a, &b).unwrap();
+/// assert!((similarity - 0.0).abs() < 1e-6);
+///
+/// // Using euclidean distance
+/// let distance = DistanceMetric::Euclidean.compute_distance(&a, &b).unwrap();
+/// assert!((distance - std::f32::consts::SQRT_2).abs() < 1e-6);
+/// ```
+///
+/// # Performance
+///
+/// All metrics use SIMD acceleration (AVX2/SSE2) when available. For maximum
+/// performance with large-scale similarity search:
+///
+/// 1. Pre-normalize vectors with [`normalize`] or [`normalize_in_place`]
+/// 2. Use [`DotProduct`](DistanceMetric::DotProduct) metric (single SIMD operation)
+/// 3. Store normalized vectors to avoid repeated normalization
+///
+/// # Future Enhancements
+///
+/// - **Serialization**: When serde is added as a dependency, this enum will
+///   support `#[serde(rename_all = "snake_case")]` for JSON/config serialization
+/// - **Batch operations**: `compute_distances_batch()` for SIMD-optimized
+///   multi-vector distance computation
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
+pub enum DistanceMetric {
+    /// Cosine similarity/distance.
+    ///
+    /// Measures the cosine of the angle between two vectors, making it
+    /// **scale-invariant** (only direction matters, not magnitude).
+    ///
+    /// - **Similarity**: Range [-1, 1] where 1 = identical direction,
+    ///   0 = orthogonal, -1 = opposite direction
+    /// - **Distance**: Computed as `1 - similarity`, range [0, 2]
+    ///
+    /// # When to Use
+    ///
+    /// - Text embeddings (word2vec, BERT, OpenAI embeddings)
+    /// - Semantic similarity where magnitude doesn't matter
+    /// - When vectors may have different scales
+    ///
+    /// # Implementation Note
+    ///
+    /// Uses [`cosine_similarity`] internally, which handles zero vectors
+    /// by returning 0.0 similarity.
+    #[default]
+    Cosine,
+
+    /// Euclidean (L2) distance.
+    ///
+    /// Measures the straight-line distance between two points in vector space.
+    /// Also known as the L2 norm of the difference vector.
+    ///
+    /// - **Distance**: Range [0, ∞) where 0 = identical vectors
+    /// - **Similarity**: Computed as `1 / (1 + distance)`, range (0, 1]
+    ///
+    /// # When to Use
+    ///
+    /// - Spatial data (coordinates, positions)
+    /// - Image feature vectors
+    /// - When absolute magnitude differences matter
+    /// - K-means clustering (uses squared Euclidean internally)
+    ///
+    /// # Implementation Note
+    ///
+    /// Uses [`euclidean_distance`] internally, which uses SIMD-accelerated
+    /// squared difference computation.
+    Euclidean,
+
+    /// Inner (dot) product.
+    ///
+    /// Computes the sum of element-wise products. For normalized vectors,
+    /// this equals cosine similarity but is faster (single SIMD operation).
+    ///
+    /// - **Raw value**: Range (-∞, ∞)
+    /// - **Similarity**: Raw dot product value (higher = more similar)
+    /// - **Distance**: Computed as `1 - dot_product`, which is meaningful
+    ///   only for normalized vectors
+    ///
+    /// # When to Use
+    ///
+    /// - **Maximum Inner Product Search (MIPS)**: When you want the highest
+    ///   dot product, not necessarily the closest vector
+    /// - **Pre-normalized vectors**: Equivalent to cosine but faster
+    /// - **Learned embeddings**: Some models are trained with dot product loss
+    ///
+    /// # Important
+    ///
+    /// For non-normalized vectors, dot product is **not** a proper distance
+    /// metric (doesn't satisfy triangle inequality). Use [`Cosine`](DistanceMetric::Cosine)
+    /// for general similarity or ensure vectors are normalized first.
+    DotProduct,
+}
+
+impl DistanceMetric {
+    /// Computes the distance between two vectors using this metric.
+    ///
+    /// Lower values indicate more similar vectors.
+    ///
+    /// # Returns
+    ///
+    /// - [`Cosine`](DistanceMetric::Cosine): `1 - cosine_similarity`, range [0, 2]
+    /// - [`Euclidean`](DistanceMetric::Euclidean): L2 distance, range [0, ∞)
+    /// - [`DotProduct`](DistanceMetric::DotProduct): `1 - dot_product` (meaningful only for normalized vectors)
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if the vectors have different lengths.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use gallifreydb::core::vector::DistanceMetric;
+    ///
+    /// let a = vec![1.0, 0.0];
+    /// let b = vec![1.0, 0.0];
+    ///
+    /// // Identical vectors have zero distance
+    /// assert!((DistanceMetric::Cosine.compute_distance(&a, &b).unwrap() - 0.0).abs() < 1e-6);
+    /// assert!((DistanceMetric::Euclidean.compute_distance(&a, &b).unwrap() - 0.0).abs() < 1e-6);
+    /// ```
+    #[inline]
+    pub fn compute_distance(&self, a: &[f32], b: &[f32]) -> Result<f32> {
+        match self {
+            DistanceMetric::Cosine => cosine_similarity(a, b).map(|sim| 1.0 - sim),
+            DistanceMetric::Euclidean => euclidean_distance(a, b),
+            DistanceMetric::DotProduct => dot_product(a, b).map(|dp| 1.0 - dp),
+        }
+    }
+
+    /// Computes the similarity between two vectors using this metric.
+    ///
+    /// Higher values indicate more similar vectors.
+    ///
+    /// # Returns
+    ///
+    /// - [`Cosine`](DistanceMetric::Cosine): Cosine similarity, range [-1, 1]
+    /// - [`Euclidean`](DistanceMetric::Euclidean): `1 / (1 + distance)`, range (0, 1]
+    /// - [`DotProduct`](DistanceMetric::DotProduct): Raw dot product, range (-∞, ∞)
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if the vectors have different lengths.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use gallifreydb::core::vector::DistanceMetric;
+    ///
+    /// let a = vec![1.0, 0.0];
+    /// let b = vec![1.0, 0.0];
+    ///
+    /// // Identical vectors have maximum similarity
+    /// assert!((DistanceMetric::Cosine.compute_similarity(&a, &b).unwrap() - 1.0).abs() < 1e-6);
+    /// assert!((DistanceMetric::Euclidean.compute_similarity(&a, &b).unwrap() - 1.0).abs() < 1e-6);
+    /// ```
+    #[inline]
+    pub fn compute_similarity(&self, a: &[f32], b: &[f32]) -> Result<f32> {
+        match self {
+            DistanceMetric::Cosine => cosine_similarity(a, b),
+            DistanceMetric::Euclidean => euclidean_distance(a, b).map(|dist| 1.0 / (1.0 + dist)),
+            DistanceMetric::DotProduct => dot_product(a, b),
+        }
+    }
+
+    /// Returns a human-readable name for this metric.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use gallifreydb::core::vector::DistanceMetric;
+    ///
+    /// assert_eq!(DistanceMetric::Cosine.name(), "cosine");
+    /// assert_eq!(DistanceMetric::Euclidean.name(), "euclidean");
+    /// assert_eq!(DistanceMetric::DotProduct.name(), "dot_product");
+    /// ```
+    #[inline]
+    pub const fn name(&self) -> &'static str {
+        match self {
+            DistanceMetric::Cosine => "cosine",
+            DistanceMetric::Euclidean => "euclidean",
+            DistanceMetric::DotProduct => "dot_product",
+        }
+    }
+
+    /// Returns whether this metric requires normalized vectors for optimal results.
+    ///
+    /// - [`Cosine`](DistanceMetric::Cosine): No (handles normalization internally)
+    /// - [`Euclidean`](DistanceMetric::Euclidean): No (works with any vectors)
+    /// - [`DotProduct`](DistanceMetric::DotProduct): Yes (otherwise not a proper similarity)
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use gallifreydb::core::vector::DistanceMetric;
+    ///
+    /// assert!(!DistanceMetric::Cosine.requires_normalized_vectors());
+    /// assert!(!DistanceMetric::Euclidean.requires_normalized_vectors());
+    /// assert!(DistanceMetric::DotProduct.requires_normalized_vectors());
+    /// ```
+    #[inline]
+    pub const fn requires_normalized_vectors(&self) -> bool {
+        matches!(self, DistanceMetric::DotProduct)
+    }
+}
+
+impl fmt::Display for DistanceMetric {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.name())
+    }
+}
