+++ src/core/vector/tests.rs
@@ -1,2526 +1,7 @@
-//! Vector utilities for GallifreyDB.
-//!
-//! This module provides types and functions for working with dense vectors
-//! (embeddings) used in semantic search and similarity operations.
-//!
-//! # Overview
-//!
-//! GallifreyDB supports storing vectors as property values on nodes via
-//! [`PropertyValue::Vector`]. This module provides the utilities needed
-//! to work with those vectors effectively:
-//!
-//! - **Type definitions**: [`VectorDimension`] for expressing vector sizes
-//! - **Similarity functions**: [`cosine_similarity`], [`cosine_similarity_normalized`]
-//! - **Distance functions**: [`euclidean_distance`], [`squared_euclidean_distance`]
-//! - **Inner product**: [`dot_product`] for pre-normalized vectors or projections
-//! - **Normalization**: [`magnitude`], [`squared_magnitude`], [`normalize`], [`normalize_in_place`], [`is_normalized`]
-//! - **Validation**: [`validate_vector`], [`check_dimensions_match`] for NaN/Inf detection and dimension checking
-//!
-//! # Usage
-//!
-//! ```rust
-//! use gallifreydb::core::vector::VectorDimension;
-//! use gallifreydb::core::PropertyValue;
-//!
-//! // Create a vector property
-//! let embedding: Vec<f32> = vec![0.1, 0.2, 0.3, 0.4];
-//! let dim = VectorDimension::new(embedding.len());
-//! let prop = PropertyValue::vector(embedding);
-//!
-//! // Access the vector
-//! if let Some(vec) = prop.as_vector() {
-//!     assert_eq!(vec.len(), dim.as_usize());
-//! }
-//! ```
-//!
-//! # Design Notes
-//!
-//! Vectors in GallifreyDB are stored as `Arc<[f32]>` within [`PropertyValue::Vector`].
-//! This design enables:
-//!
-//! - **Efficient cloning**: Multiple versions can share the same vector data
-//! - **Memory efficiency**: f32 provides good precision with half the memory of f64
-//! - **Temporal compatibility**: Unchanged vectors across versions share storage
-//!
-//! For similarity computations, vectors should typically be L2-normalized to enable
-//! fast cosine similarity via dot product.
-//!
-//! # Type Safety
-//!
-//! [`VectorDimension`] is implemented as a newtype struct rather than a type alias.
-//! This provides stronger type safety by preventing accidental interchange with
-//! other `usize` values (e.g., byte counts, array indices).
-//!
-//! # Implemented Functions
-//!
-//! - **[`cosine_similarity`]**: Measures angle between vectors, range `[-1, 1]`
-//! - **[`cosine_similarity_normalized`]**: Optimized for pre-normalized (unit) vectors
-//! - **[`euclidean_distance`]**: L2 distance between vectors
-//! - **[`squared_euclidean_distance`]**: Squared L2 distance (faster for comparisons)
-//! - **[`dot_product`]**: Inner product, useful for pre-normalized vectors
-//! - **[`magnitude`]**: L2 norm of a vector
-//! - **[`squared_magnitude`]**: Squared L2 norm (faster for comparisons)
-//! - **[`normalize`]**: Returns new unit vector with magnitude 1.0
-//! - **[`normalize_in_place`]**: Normalizes vector in place
-//! - **[`is_normalized`]**: Checks if vector has unit magnitude
-//!
-//! All functions use SIMD acceleration (AVX2/SSE2) when available.
-//!
-//! # Future Additions
-//!
-//! This module will be expanded to include:
-//!
-//! - Manhattan distance
-//! - Dimension validation helpers
-//! - Sparse vector support
-//!
-//! See `docs/VECTOR_SEARCH_DESIGN.md` for the complete design.
-
+use super::*;
 use crate::core::property::MAX_VECTOR_DIMENSIONS;
-use crate::utils::error::{Error, Result, VectorError};
-use std::fmt;
-
-// ============================================================================
-// Constants
-// ============================================================================
-
-/// Default tolerance for floating-point comparisons in normalization operations.
-///
-/// This tolerance (1e-6) is appropriate for most f32 operations where accumulated
-/// floating-point errors are expected to be small. It's used as the default for
-/// functions like [`is_normalized`] when checking if a vector has unit magnitude.
-///
-/// For stricter or looser comparisons, functions accept an explicit tolerance parameter.
-pub const NORMALIZATION_TOLERANCE: f32 = 1e-6;
-
-/// Squared magnitude threshold for detecting near-zero vectors.
-///
-/// Vectors with squared magnitude below this threshold are treated as zero vectors
-/// in normalization operations. This prevents numerical instability from denormal
-/// numbers and avoids division by very small values that could cause overflow.
-///
-/// Value: 1e-14 corresponds to magnitude ≈ 1e-7, providing safety margin for f32
-/// precision (which has ~7 significant digits). This is more conservative than
-/// 1e-20 (magnitude ≈ 1e-10) which is too close to f32's precision limits.
-const SQUARED_MAGNITUDE_THRESHOLD: f32 = 1e-14;
-
-// ============================================================================
-// Type Definitions
-// ============================================================================
-
-/// The dimension (number of elements) of a vector.
-///
-/// This is a newtype wrapper around `usize` that provides type safety when
-/// working with vector dimensions throughout the codebase. Unlike a type alias,
-/// this prevents accidental interchange with other `usize` values.
-///
-/// # Common Embedding Dimensions
-///
-/// | Model | Dimensions |
-/// |-------|------------|
-/// | OpenAI text-embedding-3-small | 1536 |
-/// | OpenAI text-embedding-3-large | 3072 |
-/// | Cohere embed-v3 | 1024 |
-/// | Sentence Transformers (all-MiniLM) | 384 |
-/// | BGE models | 768-1024 |
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::VectorDimension;
-///
-/// fn check_dimensions(vec: &[f32], expected: VectorDimension) -> bool {
-///     vec.len() == expected.as_usize()
-/// }
-///
-/// let embedding = vec![0.1f32; 384];
-/// assert!(check_dimensions(&embedding, VectorDimension::new(384)));
-/// ```
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Default)]
-pub struct VectorDimension(usize);
-
-impl VectorDimension {
-    /// Creates a new `VectorDimension` from a `usize` value.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::VectorDimension;
-    ///
-    /// let dim = VectorDimension::new(1536);
-    /// assert_eq!(dim.as_usize(), 1536);
-    /// ```
-    #[inline]
-    pub const fn new(value: usize) -> Self {
-        Self(value)
-    }
-
-    /// Returns the dimension as a `usize`.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::VectorDimension;
-    ///
-    /// let dim = VectorDimension::new(384);
-    /// let size: usize = dim.as_usize();
-    /// assert_eq!(size, 384);
-    /// ```
-    #[inline]
-    pub const fn as_usize(self) -> usize {
-        self.0
-    }
-
-    /// Returns `true` if this dimension is zero.
-    #[inline]
-    pub const fn is_zero(self) -> bool {
-        self.0 == 0
-    }
-
-    /// Returns `true` if this dimension exceeds the maximum allowed.
-    #[inline]
-    pub const fn exceeds_max(self) -> bool {
-        self.0 > MAX_VECTOR_DIMENSIONS
-    }
-}
-
-impl fmt::Display for VectorDimension {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl From<usize> for VectorDimension {
-    #[inline]
-    fn from(value: usize) -> Self {
-        Self(value)
-    }
-}
-
-impl From<VectorDimension> for usize {
-    #[inline]
-    fn from(dim: VectorDimension) -> Self {
-        dim.0
-    }
-}
-
-// ============================================================================
-// Sparse Vector Type
-// ============================================================================
-
-/// A sparse vector representation optimized for vectors with many zero values.
-///
-/// Sparse vectors store only non-zero values along with their indices, making them
-/// memory-efficient for high-dimensional vectors where most values are zero.
-/// This is particularly useful for algorithms like BM25 and SPLADE.
-///
-/// # Format
-///
-/// - `indices`: Sorted array of non-zero element positions
-/// - `values`: Corresponding non-zero values
-/// - `dimension`: Total vector dimension (including zeros)
-///
-/// # Invariants
-///
-/// The struct maintains these invariants:
-/// 1. `indices.len() == values.len()` (each index has a corresponding value)
-/// 2. `indices` are sorted in ascending order
-/// 3. All indices are `< dimension`
-/// 4. No duplicate indices
-/// 5. All values are non-zero (no stored zeros)
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::SparseVec;
-///
-/// // Sparse vector: [0.0, 1.5, 0.0, 0.0, 2.3, 0.0, 0.0, 0.8]
-/// let sparse = SparseVec::new(
-///     vec![1, 4, 7],        // indices of non-zero values
-///     vec![1.5, 2.3, 0.8],  // corresponding values
-///     8                      // total dimension
-/// ).unwrap();
-///
-/// assert_eq!(sparse.nnz(), 3);  // 3 non-zero elements
-/// assert_eq!(sparse.dimension(), 8);
-/// ```
-///
-/// # Use Cases
-///
-/// - **BM25**: Text retrieval using term frequency vectors
-/// - **SPLADE**: Sparse learned embeddings for information retrieval
-/// - **One-hot encodings**: Categorical features with single non-zero value
-/// - **TF-IDF**: Document vectors with few non-zero terms
-///
-/// # Performance
-///
-/// - Space complexity: O(nnz) where nnz = number of non-zero elements
-/// - Dense equivalent would be: O(dimension)
-/// - Memory savings can be 10-1000x for sparse data
-///
-/// # Equality and Comparison
-///
-/// While this type implements `PartialEq` for compatibility with `PropertyValue`,
-/// direct equality comparison using `==` is **not recommended** for floating-point
-/// values. NaN != NaN (IEEE 754) and floating-point precision issues can cause
-/// semantically equal vectors to compare unequal. For robust equality checks,
-/// use [`approx_eq`](Self::approx_eq) with an appropriate epsilon value instead.
-#[derive(Debug, Clone, PartialEq)]
-pub struct SparseVec {
-    /// Indices of non-zero elements (sorted, unique, all < dimension).
-    indices: Vec<u32>,
-    /// Non-zero values corresponding to indices.
-    values: Vec<f32>,
-    /// Total dimension of the vector (including zeros).
-    dimension: u32,
-}
-
-impl SparseVec {
-    /// Creates a new sparse vector.
-    ///
-    /// # Arguments
-    ///
-    /// * `indices` - Positions of non-zero elements (will be sorted if not already)
-    /// * `values` - Non-zero values corresponding to indices
-    /// * `dimension` - Total vector dimension (must be > max(indices))
-    ///
-    /// # Returns
-    ///
-    /// - `Ok(SparseVec)` if the input is valid
-    /// - `Err` if validation fails (see Errors section)
-    ///
-    /// # Errors
-    ///
-    /// Returns `VectorError::DimensionMismatch` if:
-    /// - `indices.len() != values.len()`
-    /// - Any index >= dimension
-    ///
-    /// Returns `VectorError::InvalidSparseVector` if:
-    /// - Duplicate indices found
-    /// - Zero values found (sparse vectors should only store non-zero values)
-    ///
-    /// Returns `VectorError::DimensionTooLarge` if dimension exceeds MAX_VECTOR_DIMENSIONS.
-    ///
-    /// Returns `VectorError::ContainsNaN` if any value is NaN.
-    ///
-    /// Returns `VectorError::ContainsInfinity` if any value is infinite.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::SparseVec;
-    ///
-    /// // Valid sparse vector
-    /// let sparse = SparseVec::new(vec![0, 2, 5], vec![1.0, 2.0, 3.0], 10).unwrap();
-    /// assert_eq!(sparse.nnz(), 3);
-    ///
-    /// // Error: mismatched lengths
-    /// let result = SparseVec::new(vec![0, 1], vec![1.0], 10);
-    /// assert!(result.is_err());
-    ///
-    /// // Error: index out of bounds
-    /// let result = SparseVec::new(vec![0, 10], vec![1.0, 2.0], 10);
-    /// assert!(result.is_err());
-    /// ```
-    pub fn new(mut indices: Vec<u32>, mut values: Vec<f32>, dimension: u32) -> Result<Self> {
-        // Validate dimension
-        if dimension as usize > MAX_VECTOR_DIMENSIONS {
-            return Err(VectorError::DimensionTooLarge {
-                dimension: dimension as usize,
-                max_allowed: MAX_VECTOR_DIMENSIONS,
-            }
-            .into());
-        }
-
-        // Validate lengths match
-        if indices.len() != values.len() {
-            return Err(VectorError::DimensionMismatch {
-                expected: indices.len(),
-                actual: values.len(),
-            }
-            .into());
-        }
-
-        // Validate and normalize data
-        if !indices.is_empty() {
-            // Sort by indices (if not already sorted)
-            let mut index_value_pairs: Vec<(u32, f32)> = indices.into_iter().zip(values).collect();
-            index_value_pairs.sort_by_key(|(idx, _)| *idx);
-
-            // Check for duplicates and out-of-bounds indices
-            let mut prev_idx = None;
-            for (idx, val) in &index_value_pairs {
-                // Check NaN
-                if val.is_nan() {
-                    return Err(VectorError::ContainsNaN { count: 1 }.into());
-                }
-                // Check Infinity
-                if val.is_infinite() {
-                    return Err(VectorError::ContainsInfinity { count: 1 }.into());
-                }
-                // Check for zero values (sparse vectors should not store zeros)
-                if *val == 0.0 {
-                    return Err(Error::Vector(VectorError::InvalidSparseVector {
-                        reason: "Sparse vector contains zero value".to_string(),
-                    }));
-                }
-                // Check index bounds
-                if *idx >= dimension {
-                    return Err(Error::Vector(VectorError::InvalidSparseVector {
-                        reason: format!(
-                            "Index {} is out of bounds for dimension {}",
-                            idx, dimension
-                        ),
-                    }));
-                }
-                // Check for duplicates
-                if let Some(prev) = prev_idx
-                    && prev == *idx
-                {
-                    return Err(Error::Vector(VectorError::InvalidSparseVector {
-                        reason: format!("Duplicate index {} found", idx),
-                    }));
-                }
-                prev_idx = Some(*idx);
-            }
-
-            // Unzip back to separate vectors
-            let (sorted_indices, sorted_values): (Vec<u32>, Vec<f32>) =
-                index_value_pairs.into_iter().unzip();
-            indices = sorted_indices;
-            values = sorted_values;
-        }
-
-        Ok(Self {
-            indices,
-            values,
-            dimension,
-        })
-    }
-
-    /// Returns the number of non-zero elements.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::SparseVec;
-    ///
-    /// let sparse = SparseVec::new(vec![1, 3, 5], vec![1.0, 2.0, 3.0], 10).unwrap();
-    /// assert_eq!(sparse.nnz(), 3);
-    /// ```
-    #[inline]
-    pub fn nnz(&self) -> usize {
-        self.indices.len()
-    }
-
-    /// Returns the total dimension of the vector (including zeros).
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::SparseVec;
-    ///
-    /// let sparse = SparseVec::new(vec![0], vec![1.0], 100).unwrap();
-    /// assert_eq!(sparse.dimension(), 100);
-    /// ```
-    #[inline]
-    pub fn dimension(&self) -> usize {
-        self.dimension as usize
-    }
-
-    /// Returns a slice of the non-zero indices.
-    ///
-    /// The indices are guaranteed to be sorted in ascending order.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::SparseVec;
-    ///
-    /// let sparse = SparseVec::new(vec![5, 1, 3], vec![1.0, 2.0, 3.0], 10).unwrap();
-    /// // Indices are sorted during construction
-    /// assert_eq!(sparse.indices(), &[1, 3, 5]);
-    /// ```
-    #[inline]
-    pub fn indices(&self) -> &[u32] {
-        &self.indices
-    }
-
-    /// Returns a slice of the non-zero values.
-    ///
-    /// The values correspond to the indices returned by [`indices()`](Self::indices).
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::SparseVec;
-    ///
-    /// let sparse = SparseVec::new(vec![5, 1, 3], vec![1.0, 2.0, 3.0], 10).unwrap();
-    /// let indices = sparse.indices();
-    /// let values = sparse.values();
-    /// // values[i] corresponds to indices[i]
-    /// assert_eq!(values.len(), indices.len());
-    /// ```
-    #[inline]
-    pub fn values(&self) -> &[f32] {
-        &self.values
-    }
-
-    /// Converts this sparse vector to a dense vector representation.
-    ///
-    /// Creates a new vector of length `dimension()` with zeros everywhere
-    /// except at the specified indices.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::SparseVec;
-    ///
-    /// let sparse = SparseVec::new(vec![1, 3], vec![1.5, 2.5], 5).unwrap();
-    /// let dense = sparse.to_dense();
-    /// assert_eq!(dense, vec![0.0, 1.5, 0.0, 2.5, 0.0]);
-    /// ```
-    pub fn to_dense(&self) -> Vec<f32> {
-        let mut dense = vec![0.0; self.dimension as usize];
-        for (&idx, &val) in self.indices.iter().zip(self.values.iter()) {
-            dense[idx as usize] = val;
-        }
-        dense
-    }
-
-    /// Computes the squared magnitude (L2 norm squared) of this sparse vector.
-    ///
-    /// This is more efficient than computing the full magnitude when you only
-    /// need to compare magnitudes (since sqrt is monotonic for positive values).
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::SparseVec;
-    ///
-    /// let sparse = SparseVec::new(vec![0, 1, 2], vec![1.0, 2.0, 2.0], 5).unwrap();
-    /// // magnitude² = 1² + 2² + 2² = 1 + 4 + 4 = 9
-    /// assert_eq!(sparse.squared_magnitude(), 9.0);
-    /// ```
-    pub fn squared_magnitude(&self) -> f32 {
-        self.values.iter().map(|v| v * v).sum()
-    }
-
-    /// Computes the magnitude (L2 norm) of this sparse vector.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::SparseVec;
-    ///
-    /// let sparse = SparseVec::new(vec![0, 1], vec![3.0, 4.0], 5).unwrap();
-    /// // magnitude = sqrt(3² + 4²) = sqrt(9 + 16) = 5.0
-    /// assert_eq!(sparse.magnitude(), 5.0);
-    /// ```
-    #[inline]
-    pub fn magnitude(&self) -> f32 {
-        self.squared_magnitude().sqrt()
-    }
-
-    /// Checks if this sparse vector is approximately equal to another within a tolerance.
-    ///
-    /// This method provides an epsilon-based comparison that handles floating-point
-    /// precision issues. Two sparse vectors are considered approximately equal if:
-    /// - They have the same dimension
-    /// - They have the same indices
-    /// - All corresponding values differ by less than epsilon
-    ///
-    /// # Arguments
-    ///
-    /// * `other` - The sparse vector to compare with
-    /// * `epsilon` - Maximum allowed difference for each value (typically 1e-6 for f32)
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::SparseVec;
-    ///
-    /// let a = SparseVec::new(vec![0, 2], vec![1.0, 2.0], 5).unwrap();
-    /// let b = SparseVec::new(vec![0, 2], vec![1.0000001, 2.0000001], 5).unwrap();
-    ///
-    /// // Small floating-point differences are tolerated
-    /// assert!(a.approx_eq(&b, 1e-5));
-    /// assert!(!a.approx_eq(&b, 1e-10));
-    /// ```
-    ///
-    /// # Note
-    ///
-    /// This is the **recommended way to compare sparse vectors**. While `SparseVec`
-    /// implements `PartialEq` for compatibility with `PropertyValue`, direct equality
-    /// via `==` is not recommended due to floating-point comparison concerns.
-    /// See the "Equality and Comparison" section in the type documentation for details.
-    pub fn approx_eq(&self, other: &SparseVec, epsilon: f32) -> bool {
-        self.dimension == other.dimension
-            && self.indices == other.indices
-            && self.values.len() == other.values.len()
-            && self
-                .values
-                .iter()
-                .zip(other.values.iter())
-                .all(|(a, b)| (a - b).abs() < epsilon)
-    }
-}
-
-// ============================================================================
-// Constants
-// ============================================================================
-
-/// Maximum allowed vector dimension.
-///
-/// Re-exported from [`crate::core::property::MAX_VECTOR_DIMENSIONS`] for convenience.
-/// This limit (100,000) far exceeds typical embedding sizes and exists to prevent
-/// DoS attacks via memory exhaustion during deserialization.
-pub const MAX_DIMENSION: VectorDimension = VectorDimension(MAX_VECTOR_DIMENSIONS);
-
-// ============================================================================
-// SIMD Support
-// ============================================================================
-
-/// SIMD-accelerated vector operations for x86/x86_64 platforms.
-///
-/// Uses runtime feature detection to select the best available instruction set:
-/// - AVX2 (256-bit vectors, 8 floats at a time)
-/// - SSE2 (128-bit vectors, 4 floats at a time) - baseline for x86_64
-/// - Scalar fallback for other platforms
-///
-/// # Performance Expectations
-///
-/// Compared to scalar implementation, expected speedups for vectors > 256 dimensions:
-/// - **AVX2 + FMA**: ~5-8x speedup (processes 8 floats/cycle with fused multiply-add)
-/// - **SSE2**: ~2-4x speedup (processes 4 floats/cycle)
-///
-/// For smaller vectors, SIMD overhead may reduce gains. The crossover point
-/// where SIMD becomes beneficial is typically around 16-32 dimensions.
-#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-mod simd {
-    #[cfg(target_arch = "x86")]
-    use std::arch::x86::*;
-    #[cfg(target_arch = "x86_64")]
-    use std::arch::x86_64::*;
-
-    /// Computes dot product, magnitude_a², and magnitude_b² using AVX2.
-    ///
-    /// # Safety
-    /// Caller must ensure:
-    /// 1. AVX2 and FMA are available (checked via `is_x86_feature_detected!`).
-    /// 2. `a.len() == b.len()`.
-    #[target_feature(enable = "avx2", enable = "fma")]
-    #[inline]
-    pub unsafe fn dot_and_magnitudes_avx2(a: &[f32], b: &[f32]) -> (f32, f32, f32) {
-        unsafe {
-            let len = a.len();
-            let chunks = len / 8;
-            let remainder = len % 8;
-
-            // Accumulators for 8 floats at a time
-            let mut dot_acc = _mm256_setzero_ps();
-            let mut mag_a_acc = _mm256_setzero_ps();
-            let mut mag_b_acc = _mm256_setzero_ps();
-
-            let a_ptr = a.as_ptr();
-            let b_ptr = b.as_ptr();
-
-            // Process 8 floats at a time
-            for i in 0..chunks {
-                let offset = i * 8;
-                let va = _mm256_loadu_ps(a_ptr.add(offset));
-                let vb = _mm256_loadu_ps(b_ptr.add(offset));
-
-                // Fused multiply-add for dot product and magnitudes
-                dot_acc = _mm256_fmadd_ps(va, vb, dot_acc);
-                mag_a_acc = _mm256_fmadd_ps(va, va, mag_a_acc);
-                mag_b_acc = _mm256_fmadd_ps(vb, vb, mag_b_acc);
-            }
-
-            // Horizontal sum of 256-bit vectors
-            let dot = horizontal_sum_avx(dot_acc);
-            let mag_a = horizontal_sum_avx(mag_a_acc);
-            let mag_b = horizontal_sum_avx(mag_b_acc);
-
-            // Handle remainder with safe scalar operations.
-            // Using safe indexing here as the compiler optimizes away bounds checks
-            // when the loop bound is known to be < 8 (the chunk size).
-            let mut dot_rem = 0.0f32;
-            let mut mag_a_rem = 0.0f32;
-            let mut mag_b_rem = 0.0f32;
-
-            let start = chunks * 8;
-            for i in 0..remainder {
-                let ai = a[start + i];
-                let bi = b[start + i];
-                dot_rem += ai * bi;
-                mag_a_rem += ai * ai;
-                mag_b_rem += bi * bi;
-            }
-
-            (dot + dot_rem, mag_a + mag_a_rem, mag_b + mag_b_rem)
-        }
-    }
-
-    /// Horizontal sum of 8 floats in AVX register.
-    #[target_feature(enable = "avx2")]
-    #[inline]
-    unsafe fn horizontal_sum_avx(v: __m256) -> f32 {
-        unsafe {
-            // Add high 128 bits to low 128 bits
-            let high = _mm256_extractf128_ps(v, 1);
-            let low = _mm256_castps256_ps128(v);
-            let sum128 = _mm_add_ps(high, low);
-
-            // Continue with SSE horizontal add
-            horizontal_sum_sse(sum128)
-        }
-    }
-
-    /// Computes dot product, magnitude_a², and magnitude_b² using SSE2.
-    ///
-    /// # Safety
-    /// Caller must ensure:
-    /// 1. SSE2 is available (always true on x86_64).
-    /// 2. `a.len() == b.len()`.
-    #[target_feature(enable = "sse2")]
-    #[inline]
-    pub unsafe fn dot_and_magnitudes_sse2(a: &[f32], b: &[f32]) -> (f32, f32, f32) {
-        unsafe {
-            let len = a.len();
-            let chunks = len / 4;
-            let remainder = len % 4;
-
-            // Accumulators for 4 floats at a time
-            let mut dot_acc = _mm_setzero_ps();
-            let mut mag_a_acc = _mm_setzero_ps();
-            let mut mag_b_acc = _mm_setzero_ps();
-
-            let a_ptr = a.as_ptr();
-            let b_ptr = b.as_ptr();
-
-            // Process 4 floats at a time
-            for i in 0..chunks {
-                let offset = i * 4;
-                let va = _mm_loadu_ps(a_ptr.add(offset));
-                let vb = _mm_loadu_ps(b_ptr.add(offset));
-
-                // Multiply and accumulate
-                dot_acc = _mm_add_ps(dot_acc, _mm_mul_ps(va, vb));
-                mag_a_acc = _mm_add_ps(mag_a_acc, _mm_mul_ps(va, va));
-                mag_b_acc = _mm_add_ps(mag_b_acc, _mm_mul_ps(vb, vb));
-            }
-
-            // Horizontal sum of 128-bit vectors
-            let dot = horizontal_sum_sse(dot_acc);
-            let mag_a = horizontal_sum_sse(mag_a_acc);
-            let mag_b = horizontal_sum_sse(mag_b_acc);
-
-            // Handle remainder with safe scalar operations.
-            // Using safe indexing here as the compiler optimizes away bounds checks
-            // when the loop bound is known to be < 4 (the chunk size).
-            let mut dot_rem = 0.0f32;
-            let mut mag_a_rem = 0.0f32;
-            let mut mag_b_rem = 0.0f32;
-
-            let start = chunks * 4;
-            for i in 0..remainder {
-                let ai = a[start + i];
-                let bi = b[start + i];
-                dot_rem += ai * bi;
-                mag_a_rem += ai * ai;
-                mag_b_rem += bi * bi;
-            }
-
-            (dot + dot_rem, mag_a + mag_a_rem, mag_b + mag_b_rem)
-        }
-    }
-
-    /// Horizontal sum of 4 floats in SSE register.
-    #[target_feature(enable = "sse2")]
-    #[inline]
-    unsafe fn horizontal_sum_sse(v: __m128) -> f32 {
-        // Sum pairs: [a+c, b+d, a+c, b+d]
-        // SAFETY: We have #[target_feature(enable = "sse2")] so these intrinsics are safe
-        let shuf = _mm_shuffle_ps(v, v, 0b10_11_00_01);
-        let sum1 = _mm_add_ps(v, shuf);
-        // Sum final pair
-        let shuf2 = _mm_shuffle_ps(sum1, sum1, 0b00_00_11_10);
-        let sum2 = _mm_add_ps(sum1, shuf2);
-        _mm_cvtss_f32(sum2)
-    }
-
-    /// Computes dot product using AVX2.
-    ///
-    /// This is a dedicated dot-product-only function, more efficient than
-    /// `dot_and_magnitudes_avx2` when magnitudes aren't needed.
-    ///
-    /// # Safety
-    /// Caller must ensure:
-    /// 1. AVX2 and FMA are available (checked via `is_x86_feature_detected!`).
-    /// 2. `a.len() == b.len()`.
-    #[target_feature(enable = "avx2", enable = "fma")]
-    #[inline]
-    pub unsafe fn dot_product_avx2(a: &[f32], b: &[f32]) -> f32 {
-        // SAFETY: The unsafe block is required by the `unsafe_op_in_unsafe_fn` lint.
-        // The caller guarantees AVX2 and FMA are available via runtime feature detection.
-        unsafe {
-            let a_chunks = a.chunks_exact(8);
-            let b_chunks = b.chunks_exact(8);
-            let a_rem = a_chunks.remainder();
-            let b_rem = b_chunks.remainder();
-
-            // Accumulator for 8 floats at a time
-            let mut acc = _mm256_setzero_ps();
-
-            // Process 8 floats at a time
-            for (va_chunk, vb_chunk) in a_chunks.zip(b_chunks) {
-                let va = _mm256_loadu_ps(va_chunk.as_ptr());
-                let vb = _mm256_loadu_ps(vb_chunk.as_ptr());
-
-                // Fused multiply-add: acc = va * vb + acc
-                acc = _mm256_fmadd_ps(va, vb, acc);
-            }
-
-            // Horizontal sum of 256-bit vector
-            let mut sum = horizontal_sum_avx(acc);
-
-            // Handle remainder with scalar operations
-            for (&va, &vb) in a_rem.iter().zip(b_rem) {
-                sum += va * vb;
-            }
-
-            sum
-        }
-    }
-
-    /// Computes dot product using SSE2.
-    ///
-    /// This is a dedicated dot-product-only function, more efficient than
-    /// `dot_and_magnitudes_sse2` when magnitudes aren't needed.
-    ///
-    /// # Safety
-    /// Caller must ensure:
-    /// 1. SSE2 is available (always true on x86_64).
-    /// 2. `a.len() == b.len()`.
-    #[target_feature(enable = "sse2")]
-    #[inline]
-    pub unsafe fn dot_product_sse2(a: &[f32], b: &[f32]) -> f32 {
-        // SAFETY: The unsafe block is required by the `unsafe_op_in_unsafe_fn` lint.
-        // The caller guarantees SSE2 is available via runtime feature detection.
-        unsafe {
-            let a_chunks = a.chunks_exact(4);
-            let b_chunks = b.chunks_exact(4);
-            let a_rem = a_chunks.remainder();
-            let b_rem = b_chunks.remainder();
-
-            // Accumulator for 4 floats at a time
-            let mut acc = _mm_setzero_ps();
-
-            // Process 4 floats at a time
-            for (va_chunk, vb_chunk) in a_chunks.zip(b_chunks) {
-                let va = _mm_loadu_ps(va_chunk.as_ptr());
-                let vb = _mm_loadu_ps(vb_chunk.as_ptr());
-
-                // Multiply and accumulate
-                acc = _mm_add_ps(acc, _mm_mul_ps(va, vb));
-            }
-
-            // Horizontal sum of 128-bit vector
-            let mut sum = horizontal_sum_sse(acc);
-
-            // Handle remainder with scalar operations
-            for (&va, &vb) in a_rem.iter().zip(b_rem) {
-                sum += va * vb;
-            }
-
-            sum
-        }
-    }
-
-    /// Computes sum of squared differences using AVX2.
-    ///
-    /// # Safety
-    /// Caller must ensure:
-    /// 1. AVX2 and FMA are available (checked via `is_x86_feature_detected!`).
-    /// 2. `a.len() == b.len()`.
-    #[target_feature(enable = "avx2", enable = "fma")]
-    #[inline]
-    pub unsafe fn squared_diff_sum_avx2(a: &[f32], b: &[f32]) -> f32 {
-        // SAFETY: The unsafe block is required by the `unsafe_op_in_unsafe_fn` lint.
-        // All unsafe operations within this unsafe fn must still be in an unsafe block.
-        // The caller guarantees AVX2 and FMA are available via runtime feature detection.
-        unsafe {
-            let len = a.len();
-            let chunks = len / 8;
-            let remainder = len % 8;
-
-            // Accumulator for 8 floats at a time
-            let mut acc = _mm256_setzero_ps();
-
-            let a_ptr = a.as_ptr();
-            let b_ptr = b.as_ptr();
-
-            // Process 8 floats at a time
-            for i in 0..chunks {
-                let offset = i * 8;
-                let va = _mm256_loadu_ps(a_ptr.add(offset));
-                let vb = _mm256_loadu_ps(b_ptr.add(offset));
-
-                // Compute difference
-                let diff = _mm256_sub_ps(va, vb);
-
-                // Square and accumulate using FMA: acc = diff * diff + acc
-                acc = _mm256_fmadd_ps(diff, diff, acc);
-            }
-
-            // Horizontal sum of 256-bit vector
-            let mut sum = horizontal_sum_avx(acc);
-
-            // Handle remainder with scalar operations
-            let start = chunks * 8;
-            for i in 0..remainder {
-                let diff = a[start + i] - b[start + i];
-                sum += diff * diff;
-            }
-
-            sum
-        }
-    }
-
-    /// Computes sum of squared differences using SSE2.
-    ///
-    /// # Safety
-    /// Caller must ensure:
-    /// 1. SSE2 is available (always true on x86_64).
-    /// 2. `a.len() == b.len()`.
-    #[target_feature(enable = "sse2")]
-    #[inline]
-    pub unsafe fn squared_diff_sum_sse2(a: &[f32], b: &[f32]) -> f32 {
-        // SAFETY: The unsafe block is required by the `unsafe_op_in_unsafe_fn` lint.
-        // All unsafe operations within this unsafe fn must still be in an unsafe block.
-        // The caller guarantees SSE2 is available via runtime feature detection.
-        unsafe {
-            let len = a.len();
-            let chunks = len / 4;
-            let remainder = len % 4;
-
-            // Accumulator for 4 floats at a time
-            let mut acc = _mm_setzero_ps();
-
-            let a_ptr = a.as_ptr();
-            let b_ptr = b.as_ptr();
-
-            // Process 4 floats at a time
-            for i in 0..chunks {
-                let offset = i * 4;
-                let va = _mm_loadu_ps(a_ptr.add(offset));
-                let vb = _mm_loadu_ps(b_ptr.add(offset));
-
-                // Compute difference
-                let diff = _mm_sub_ps(va, vb);
-
-                // Square and accumulate
-                acc = _mm_add_ps(acc, _mm_mul_ps(diff, diff));
-            }
-
-            // Horizontal sum of 128-bit vector
-            let mut sum = horizontal_sum_sse(acc);
-
-            // Handle remainder with scalar operations
-            let start = chunks * 4;
-            for i in 0..remainder {
-                let diff = a[start + i] - b[start + i];
-                sum += diff * diff;
-            }
-
-            sum
-        }
-    }
-
-    /// Scales a vector in place by a scalar using AVX2.
-    ///
-    /// # Safety
-    /// Caller must ensure AVX2 is available (checked via `is_x86_feature_detected!`).
-    #[target_feature(enable = "avx2")]
-    #[inline]
-    pub unsafe fn scale_in_place_avx2(v: &mut [f32], scalar: f32) {
-        // SAFETY: This unsafe block is required by the `unsafe_op_in_unsafe_fn` lint.
-        // - The caller guarantees AVX2 is available via runtime feature detection.
-        // - Pointer operations are safe because:
-        //   1. chunks_exact_mut(8) guarantees each chunk has exactly 8 f32 elements
-        //   2. chunk.as_ptr() returns a valid pointer to the chunk's contiguous memory
-        //   3. chunk.as_mut_ptr() returns a valid mutable pointer to the same memory
-        //   4. _mm256_loadu_ps/_mm256_storeu_ps handle unaligned access safely
-        //   5. The slice owns the memory, so no aliasing occurs
-        unsafe {
-            let scalar_vec = _mm256_set1_ps(scalar);
-            let mut chunks = v.chunks_exact_mut(8);
-
-            // Process 8 floats at a time
-            for chunk in chunks.by_ref() {
-                let va = _mm256_loadu_ps(chunk.as_ptr());
-                let result = _mm256_mul_ps(va, scalar_vec);
-                _mm256_storeu_ps(chunk.as_mut_ptr(), result);
-            }
-
-            // Handle remainder with scalar operations
-            for x in chunks.into_remainder() {
-                *x *= scalar;
-            }
-        }
-    }
-
-    /// Scales a vector in place by a scalar using SSE2.
-    ///
-    /// # Safety
-    /// Caller must ensure SSE2 is available (always true on x86_64).
-    #[target_feature(enable = "sse2")]
-    #[inline]
-    pub unsafe fn scale_in_place_sse2(v: &mut [f32], scalar: f32) {
-        // SAFETY: This unsafe block is required by the `unsafe_op_in_unsafe_fn` lint.
-        // - The caller guarantees SSE2 is available via runtime feature detection.
-        // - Pointer operations are safe because:
-        //   1. chunks_exact_mut(4) guarantees each chunk has exactly 4 f32 elements
-        //   2. chunk.as_ptr() returns a valid pointer to the chunk's contiguous memory
-        //   3. chunk.as_mut_ptr() returns a valid mutable pointer to the same memory
-        //   4. _mm_loadu_ps/_mm_storeu_ps handle unaligned access safely
-        //   5. The slice owns the memory, so no aliasing occurs
-        unsafe {
-            let scalar_vec = _mm_set1_ps(scalar);
-            let mut chunks = v.chunks_exact_mut(4);
-
-            // Process 4 floats at a time
-            for chunk in chunks.by_ref() {
-                let va = _mm_loadu_ps(chunk.as_ptr());
-                let result = _mm_mul_ps(va, scalar_vec);
-                _mm_storeu_ps(chunk.as_mut_ptr(), result);
-            }
-
-            // Handle remainder with scalar operations
-            for x in chunks.into_remainder() {
-                *x *= scalar;
-            }
-        }
-    }
-}
-
-/// Scalar fallback for computing dot product and magnitudes.
-///
-/// Used on non-x86 platforms or ancient x86 CPUs without SSE2.
-#[inline]
-#[cfg_attr(
-    all(any(target_arch = "x86", target_arch = "x86_64"), not(miri)),
-    allow(dead_code)
-)]
-fn dot_and_magnitudes_scalar(a: &[f32], b: &[f32]) -> (f32, f32, f32) {
-    a.iter().zip(b.iter()).fold(
-        (0.0f32, 0.0f32, 0.0f32),
-        |(dot, mag_a, mag_b), (&ai, &bi)| (dot + ai * bi, mag_a + ai * ai, mag_b + bi * bi),
-    )
-}
-
-/// Scalar fallback for computing sum of squared differences.
-///
-/// Used on non-x86 platforms or ancient x86 CPUs without SSE2.
-#[inline]
-#[cfg_attr(
-    all(any(target_arch = "x86", target_arch = "x86_64"), not(miri)),
-    allow(dead_code)
-)]
-fn squared_diff_sum_scalar(a: &[f32], b: &[f32]) -> f32 {
-    a.iter()
-        .zip(b.iter())
-        .map(|(&ai, &bi)| {
-            let diff = ai - bi;
-            diff * diff
-        })
-        .sum()
-}
-
-/// Scalar fallback for computing dot product.
-///
-/// Used on non-x86 platforms or ancient x86 CPUs without SSE2.
-#[inline]
-#[cfg_attr(
-    all(any(target_arch = "x86", target_arch = "x86_64"), not(miri)),
-    allow(dead_code)
-)]
-fn dot_product_scalar(a: &[f32], b: &[f32]) -> f32 {
-    a.iter().zip(b.iter()).map(|(&ai, &bi)| ai * bi).sum()
-}
-
-/// Scalar fallback for scaling a vector in place.
-///
-/// Used on non-x86 platforms or ancient x86 CPUs without SSE2.
-#[inline]
-#[cfg_attr(
-    all(any(target_arch = "x86", target_arch = "x86_64"), not(miri)),
-    allow(dead_code)
-)]
-fn scale_in_place_scalar(v: &mut [f32], scalar: f32) {
-    for x in v.iter_mut() {
-        *x *= scalar;
-    }
-}
-
-/// Scales a vector in place using the best available SIMD instructions.
-///
-/// Uses runtime feature detection to select:
-/// - AVX2 on x86/x86_64 when available
-/// - SSE2 on x86/x86_64 as fallback (baseline for x86_64)
-/// - Scalar implementation on other platforms
-#[inline(always)]
-fn scale_in_place(v: &mut [f32], scalar: f32) {
-    #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-    {
-        // Use runtime detection for best available instruction set.
-        if is_x86_feature_detected!("avx2") {
-            // SAFETY: We just verified AVX2 is available.
-            unsafe {
-                simd::scale_in_place_avx2(v, scalar);
-            }
-            return;
-        }
-        if is_x86_feature_detected!("sse2") {
-            // SAFETY: We just verified SSE2 is available. SSE2 is a baseline
-            // requirement for x86_64, so this check is mainly for 32-bit x86.
-            unsafe {
-                simd::scale_in_place_sse2(v, scalar);
-            }
-            return;
-        }
-    }
-
-    // Fallback for non-x86 platforms or x86 CPUs without SSE2.
-    scale_in_place_scalar(v, scalar);
-}
-
-/// Computes dot product and both squared magnitudes using the best available
-/// SIMD instructions.
-///
-/// Uses runtime feature detection to select:
-/// - AVX2 with FMA on x86/x86_64 when available
-/// - SSE2 on x86/x86_64 as fallback (baseline for x86_64)
-/// - Scalar implementation on other platforms
-#[inline]
-fn dot_and_magnitudes(a: &[f32], b: &[f32]) -> (f32, f32, f32) {
-    debug_assert_eq!(a.len(), b.len());
-    #[cfg(target_arch = "x86_64")]
-    {
-        // Use runtime detection for best available instruction set
-        if is_x86_feature_detected!("avx2") && is_x86_feature_detected!("fma") {
-            // SAFETY: We just verified AVX2 and FMA are available
-            return unsafe { simd::dot_and_magnitudes_avx2(a, b) };
-        }
-
-        // SAFETY: SSE2 is always available on x86_64 (baseline requirement)
-        unsafe { simd::dot_and_magnitudes_sse2(a, b) }
-    }
-
-    #[cfg(target_arch = "x86")]
-    {
-        // Use runtime detection for best available instruction set
-        if is_x86_feature_detected!("avx2") && is_x86_feature_detected!("fma") {
-            // SAFETY: We just verified AVX2 and FMA are available
-            return unsafe { simd::dot_and_magnitudes_avx2(a, b) };
-        }
-
-        if is_x86_feature_detected!("sse2") {
-            // SAFETY: We just verified SSE2 is available
-            return unsafe { simd::dot_and_magnitudes_sse2(a, b) };
-        }
-
-        // Fall through to scalar on ancient x86 without SSE2
-        return dot_and_magnitudes_scalar(a, b);
-    }
-
-    // Fallback for non-x86 platforms
-    #[cfg(not(any(target_arch = "x86", target_arch = "x86_64")))]
-    dot_and_magnitudes_scalar(a, b)
-}
-
-/// Computes sum of squared differences using the best available SIMD instructions.
-///
-/// Uses runtime feature detection to select:
-/// - AVX2 with FMA on x86/x86_64 when available
-/// - SSE2 on x86/x86_64 as fallback (baseline for x86_64)
-/// - Scalar implementation on other platforms
-#[inline]
-fn squared_diff_sum(a: &[f32], b: &[f32]) -> f32 {
-    debug_assert_eq!(a.len(), b.len());
-    #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-    {
-        // Use runtime detection for best available instruction set.
-        // The order of checks is from most to least performant.
-        if is_x86_feature_detected!("avx2") && is_x86_feature_detected!("fma") {
-            // SAFETY: We just verified AVX2 and FMA are available.
-            return unsafe { simd::squared_diff_sum_avx2(a, b) };
-        }
-        if is_x86_feature_detected!("sse2") {
-            // SAFETY: We just verified SSE2 is available. SSE2 is a baseline
-            // requirement for x86_64, so this check is mainly for 32-bit x86.
-            return unsafe { simd::squared_diff_sum_sse2(a, b) };
-        }
-    }
-
-    // Fallback for non-x86 platforms or x86 CPUs without SSE2.
-    squared_diff_sum_scalar(a, b)
-}
-
-/// Computes dot product using the best available SIMD instructions.
-///
-/// Uses runtime feature detection to select:
-/// - AVX2 with FMA on x86/x86_64 when available
-/// - SSE2 on x86/x86_64 as fallback (baseline for x86_64)
-/// - Scalar implementation on other platforms
-#[inline]
-fn dot_product_sum(a: &[f32], b: &[f32]) -> f32 {
-    debug_assert_eq!(a.len(), b.len());
-    #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-    {
-        // Use runtime detection for best available instruction set.
-        // The order of checks is from most to least performant.
-        if is_x86_feature_detected!("avx2") && is_x86_feature_detected!("fma") {
-            // SAFETY: We just verified AVX2 and FMA are available.
-            return unsafe { simd::dot_product_avx2(a, b) };
-        }
-        if is_x86_feature_detected!("sse2") {
-            // SAFETY: We just verified SSE2 is available. SSE2 is a baseline
-            // requirement for x86_64, so this check is mainly for 32-bit x86.
-            return unsafe { simd::dot_product_sse2(a, b) };
-        }
-    }
-
-    // Fallback for non-x86 platforms or x86 CPUs without SSE2.
-    dot_product_scalar(a, b)
-}
-
-// ============================================================================
-// Similarity Functions
-// ============================================================================
-
-/// Computes the cosine similarity between two vectors.
-///
-/// Cosine similarity measures the cosine of the angle between two vectors,
-/// returning a value in the range `[-1.0, 1.0]`:
-/// - `1.0`: Vectors point in the same direction (identical orientation)
-/// - `0.0`: Vectors are orthogonal (perpendicular)
-/// - `-1.0`: Vectors point in opposite directions
-///
-/// # Formula
-///
-/// ```text
-/// cosine_similarity(a, b) = (a · b) / (||a|| × ||b||)
-/// ```
-///
-/// where `a · b` is the dot product and `||a||` is the L2 norm (magnitude).
-///
-/// # Arguments
-///
-/// * `a` - First vector
-/// * `b` - Second vector (must have the same length as `a`)
-///
-/// # Returns
-///
-/// * `Ok(f32)` - The cosine similarity in range `[-1.0, 1.0]`
-/// * `Err` - If vectors have different dimensions
-///
-/// # Special Cases
-///
-/// - If either vector is a zero vector (all zeros), returns `0.0`
-/// - NaN values in input will propagate to the output (result will be NaN)
-/// - Inf values in input will propagate (may result in NaN depending on combination)
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::cosine_similarity;
-///
-/// // Identical vectors have similarity 1.0
-/// let a = vec![1.0, 0.0, 0.0];
-/// let b = vec![1.0, 0.0, 0.0];
-/// let sim = cosine_similarity(&a, &b).unwrap();
-/// assert!((sim - 1.0).abs() < 1e-6);
-///
-/// // Orthogonal vectors have similarity 0.0
-/// let a = vec![1.0, 0.0];
-/// let b = vec![0.0, 1.0];
-/// let sim = cosine_similarity(&a, &b).unwrap();
-/// assert!(sim.abs() < 1e-6);
-///
-/// // Opposite vectors have similarity -1.0
-/// let a = vec![1.0, 0.0];
-/// let b = vec![-1.0, 0.0];
-/// let sim = cosine_similarity(&a, &b).unwrap();
-/// assert!((sim + 1.0).abs() < 1e-6);
-/// ```
-///
-/// # Performance
-///
-/// This implementation uses SIMD acceleration when available:
-/// - **AVX2 + FMA**: Processes 8 floats at a time with fused multiply-add
-/// - **SSE2**: Processes 4 floats at a time (baseline for x86_64)
-/// - **Scalar**: Fallback for other platforms
-///
-/// All variants use a single-pass algorithm that computes the dot product
-/// and both magnitudes simultaneously for better cache efficiency.
-///
-/// # Numerical Precision
-///
-/// This implementation uses standard f32 accumulation, which provides sufficient
-/// precision for typical embedding use cases (dimensions up to ~10,000 with values
-/// in the range [-100, 100]). For these cases, relative error is typically < 1e-5.
-///
-/// **Precision characteristics:**
-/// - Typical embeddings (normalized, dim ≤ 4096): Excellent precision (< 1e-6 error)
-/// - Large vectors (dim > 10,000): May accumulate ~1e-4 relative error
-/// - Extreme magnitudes (values > 1e6): Consider normalizing inputs first
-///
-/// For applications requiring higher precision with extreme values, consider:
-/// 1. Normalizing vectors to unit length before comparison
-/// 2. Using f64 vectors with a custom implementation
-/// 3. Implementing Kahan summation (not included due to performance overhead)
-///
-/// The result is always clamped to `[-1.0, 1.0]` to handle minor floating-point
-/// inaccuracies that could produce values slightly outside this range.
-pub fn cosine_similarity(a: &[f32], b: &[f32]) -> Result<f32> {
-    check_dimensions_match(a, b)?;
-
-    // Handle empty vectors
-    if a.is_empty() {
-        return Ok(0.0);
-    }
-
-    // Use SIMD-accelerated computation when available
-    let (dot, mag_a_sq, mag_b_sq) = dot_and_magnitudes(a, b);
-
-    let magnitude = (mag_a_sq * mag_b_sq).sqrt();
-
-    // Handle zero vectors
-    if magnitude == 0.0 {
-        return Ok(0.0);
-    }
-
-    // Compute the raw result before clamping
-    let result = dot / magnitude;
-
-    // Debug assertion to detect if clamping is hiding a significant numerical issue.
-    // For correctly computed cosine similarity, values should only exceed [-1, 1]
-    // by at most machine epsilon (~1e-7 for f32). Values exceeding by more than
-    // 1e-5 may indicate a bug in the SIMD implementation or extreme input values.
-    debug_assert!(
-        result.is_nan() || result.abs() <= 1.0 + 1e-5,
-        "Cosine similarity {} out of valid range before clamping. \
-         This may indicate numerical issues with the input vectors.",
-        result
-    );
-
-    // Clamp to handle minor floating-point inaccuracies that could produce
-    // values slightly outside [-1.0, 1.0]
-    Ok(result.clamp(-1.0, 1.0))
-}
-
-/// Computes cosine similarity between pre-normalized (unit) vectors.
-///
-/// This is an optimized version of [`cosine_similarity`] for vectors that have
-/// already been L2-normalized (i.e., `||a|| = ||b|| = 1.0`). Since the magnitudes
-/// are known to be 1.0, this function skips the magnitude computation entirely
-/// and simply computes the dot product.
-///
-/// # Performance
-///
-/// This function provides approximately **2x speedup** compared to the general
-/// [`cosine_similarity`] function because it:
-/// - Skips computing `||a||²` and `||b||²`
-/// - Skips the `sqrt()` call for the magnitude product
-/// - Skips the division (or rather, divides by 1.0 implicitly)
-///
-/// # Arguments
-///
-/// * `a` - First unit vector (must be L2-normalized: `||a|| = 1.0`)
-/// * `b` - Second unit vector (must be L2-normalized: `||b|| = 1.0`)
-///
-/// # Returns
-///
-/// * `Ok(f32)` - The cosine similarity (equivalent to dot product for unit vectors)
-/// * `Err` - If vectors have different dimensions
-///
-/// # Panics (Debug Mode)
-///
-/// In debug builds, this function asserts that both vectors are approximately
-/// unit length. If a vector's magnitude differs from 1.0 by more than 1e-4,
-/// the assertion will fail.
-///
-/// # Safety Contract
-///
-/// The caller **must ensure** that both vectors are L2-normalized. If this
-/// precondition is violated:
-/// - Results will be mathematically incorrect
-/// - The debug assertion will catch this in debug builds
-/// - Release builds will silently produce wrong results
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::cosine_similarity_normalized;
-///
-/// // Pre-normalize vectors
-/// let a = vec![1.0, 0.0, 0.0];  // Already unit length
-/// let b_raw = vec![1.0, 1.0, 0.0];
-/// let b_mag = (b_raw.iter().map(|x| x * x).sum::<f32>()).sqrt();
-/// let b: Vec<f32> = b_raw.iter().map(|x| x / b_mag).collect();
-///
-/// let sim = cosine_similarity_normalized(&a, &b).unwrap();
-/// // sim ≈ cos(45°) ≈ 0.707
-/// assert!((sim - 0.707).abs() < 0.01);
-/// ```
-///
-/// # When to Use
-///
-/// Use this function when:
-/// - You pre-normalize vectors at ingestion time (common practice)
-/// - You're doing many similarity comparisons against the same query vector
-/// - Performance is critical and you can guarantee unit vectors
-///
-/// Use [`cosine_similarity`] instead when:
-/// - Vectors may not be normalized
-/// - You're unsure about vector magnitudes
-/// - Correctness is more important than performance
-#[inline]
-pub fn cosine_similarity_normalized(a: &[f32], b: &[f32]) -> Result<f32> {
-    check_dimensions_match(a, b)?;
-
-    // Handle empty vectors
-    if a.is_empty() {
-        return Ok(0.0);
-    }
-
-    // Debug assertions to verify the precondition that vectors are normalized
-    #[cfg(debug_assertions)]
-    {
-        let mag_a_sq: f32 = a.iter().map(|x| x * x).sum();
-        let mag_b_sq: f32 = b.iter().map(|x| x * x).sum();
-
-        debug_assert!(
-            (mag_a_sq - 1.0).abs() < 1e-4,
-            "First vector is not unit length: ||a||² = {} (expected 1.0). \
-             Use cosine_similarity() for non-normalized vectors.",
-            mag_a_sq
-        );
-        debug_assert!(
-            (mag_b_sq - 1.0).abs() < 1e-4,
-            "Second vector is not unit length: ||b||² = {} (expected 1.0). \
-             Use cosine_similarity() for non-normalized vectors.",
-            mag_b_sq
-        );
-    }
-
-    // For unit vectors, cosine similarity = dot product
-    // We reuse the SIMD infrastructure but only need the dot product
-    let dot = dot_product_sum(a, b);
-
-    // Clamp to handle floating-point inaccuracies
-    Ok(dot.clamp(-1.0, 1.0))
-}
-
-// ============================================================================
-// Distance Functions
-// ============================================================================
-
-/// Computes the squared Euclidean distance between two vectors.
-///
-/// The squared Euclidean distance is the sum of squared differences between
-/// corresponding elements. This is often preferred over [`euclidean_distance`]
-/// for comparisons because it avoids the expensive square root operation while
-/// preserving ordering (if `d²(a,b) < d²(a,c)` then `d(a,b) < d(a,c)`).
-///
-/// # Formula
-///
-/// ```text
-/// squared_euclidean_distance(a, b) = Σ(aᵢ - bᵢ)²
-/// ```
-///
-/// # Arguments
-///
-/// * `a` - First vector
-/// * `b` - Second vector (must have the same length as `a`)
-///
-/// # Returns
-///
-/// * `Ok(f32)` - The squared Euclidean distance (always non-negative)
-/// * `Err` - If vectors have different dimensions
-///
-/// # When to Use
-///
-/// Use squared distance instead of regular distance when:
-/// - Comparing distances (finding nearest neighbors)
-/// - Performance is critical
-/// - The actual distance value is not needed
-///
-/// Use [`euclidean_distance`] when:
-/// - You need the actual distance value
-/// - Combining with other non-squared metrics
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::squared_euclidean_distance;
-///
-/// // Distance from origin
-/// let a = vec![3.0, 4.0];
-/// let b = vec![0.0, 0.0];
-/// let dist_sq = squared_euclidean_distance(&a, &b).unwrap();
-/// assert!((dist_sq - 25.0).abs() < 1e-6); // 3² + 4² = 25
-///
-/// // Use for comparison without sqrt overhead
-/// let c = vec![1.0, 1.0];
-/// let dist_bc_sq = squared_euclidean_distance(&b, &c).unwrap();
-/// // dist_bc < dist_ab because dist_bc_sq < dist_ab_sq
-/// assert!(dist_bc_sq < dist_sq);
-/// ```
-///
-/// # Performance
-///
-/// This implementation uses SIMD acceleration when available:
-/// - **AVX2 + FMA**: Processes 8 floats at a time with fused multiply-add
-/// - **SSE2**: Processes 4 floats at a time (baseline for x86_64)
-/// - **Scalar**: Fallback for other platforms
-#[inline]
-pub fn squared_euclidean_distance(a: &[f32], b: &[f32]) -> Result<f32> {
-    check_dimensions_match(a, b)?;
-
-    // Handle empty vectors
-    if a.is_empty() {
-        return Ok(0.0);
-    }
-
-    // Use SIMD-accelerated computation when available
-    Ok(squared_diff_sum(a, b))
-}
-
-/// Computes the Euclidean distance between two vectors.
-///
-/// The Euclidean distance (also known as L2 distance) measures the "straight-line"
-/// distance between two points in Euclidean space. It is the most common distance
-/// metric used in machine learning and data science.
-///
-/// # Formula
-///
-/// ```text
-/// euclidean_distance(a, b) = √(Σ(aᵢ - bᵢ)²)
-/// ```
-///
-/// # Arguments
-///
-/// * `a` - First vector
-/// * `b` - Second vector (must have the same length as `a`)
-///
-/// # Returns
-///
-/// * `Ok(f32)` - The Euclidean distance (always non-negative)
-/// * `Err` - If vectors have different dimensions
-///
-/// # Performance Note
-///
-/// If you only need to compare distances (e.g., finding the k nearest neighbors),
-/// consider using [`squared_euclidean_distance`] instead, as it avoids the square
-/// root operation while preserving distance ordering.
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::euclidean_distance;
-///
-/// // Classic 3-4-5 right triangle
-/// let a = vec![0.0, 0.0];
-/// let b = vec![3.0, 4.0];
-/// let dist = euclidean_distance(&a, &b).unwrap();
-/// assert!((dist - 5.0).abs() < 1e-6);
-///
-/// // Same point has distance 0
-/// let c = vec![1.0, 2.0, 3.0];
-/// let dist_same = euclidean_distance(&c, &c).unwrap();
-/// assert!(dist_same.abs() < 1e-6);
-/// ```
-///
-/// # Performance
-///
-/// This implementation uses SIMD acceleration when available:
-/// - **AVX2 + FMA**: Processes 8 floats at a time with fused multiply-add
-/// - **SSE2**: Processes 4 floats at a time (baseline for x86_64)
-/// - **Scalar**: Fallback for other platforms
-///
-/// The square root is computed after the SIMD-accelerated sum.
-#[inline]
-pub fn euclidean_distance(a: &[f32], b: &[f32]) -> Result<f32> {
-    squared_euclidean_distance(a, b).map(|sq| sq.sqrt())
-}
-
-// ============================================================================
-// Dot Product
-// ============================================================================
-
-/// Computes the dot product (inner product) of two vectors.
-///
-/// The dot product is the sum of element-wise products: `Σ(aᵢ × bᵢ)`.
-/// It is a fundamental operation used in:
-/// - Cosine similarity (when vectors are normalized)
-/// - Linear algebra operations
-/// - Neural network computations
-/// - Projection calculations
-///
-/// # Formula
-///
-/// ```text
-/// dot_product(a, b) = Σ(aᵢ × bᵢ) = a₀×b₀ + a₁×b₁ + ... + aₙ×bₙ
-/// ```
-///
-/// # Arguments
-///
-/// * `a` - First vector
-/// * `b` - Second vector (must have the same length as `a`)
-///
-/// # Returns
-///
-/// * `Ok(f32)` - The dot product (can be positive, negative, or zero)
-/// * `Err` - If vectors have different dimensions
-///
-/// # Properties
-///
-/// - **Commutativity**: `dot(a, b) = dot(b, a)`
-/// - **Self dot product**: `dot(a, a) = ||a||²` (squared magnitude)
-/// - **Orthogonal vectors**: `dot(a, b) = 0` when vectors are perpendicular
-/// - **Parallel vectors**: `dot(a, b) = ||a|| × ||b||` (same direction)
-///   or `-||a|| × ||b||` (opposite direction)
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::dot_product;
-///
-/// // Basic dot product
-/// let a = vec![1.0, 2.0, 3.0];
-/// let b = vec![4.0, 5.0, 6.0];
-/// let result = dot_product(&a, &b).unwrap();
-/// // 1×4 + 2×5 + 3×6 = 4 + 10 + 18 = 32
-/// assert!((result - 32.0).abs() < 1e-6);
-///
-/// // Self dot product equals squared magnitude
-/// let v = vec![3.0, 4.0];
-/// let self_dot = dot_product(&v, &v).unwrap();
-/// assert!((self_dot - 25.0).abs() < 1e-6); // 3² + 4² = 25
-///
-/// // Orthogonal vectors
-/// let x = vec![1.0, 0.0];
-/// let y = vec![0.0, 1.0];
-/// let ortho = dot_product(&x, &y).unwrap();
-/// assert!(ortho.abs() < 1e-6);
-/// ```
-///
-/// # Performance
-///
-/// This implementation uses SIMD acceleration when available:
-/// - **AVX2 + FMA**: Processes 8 floats at a time with fused multiply-add
-/// - **SSE2**: Processes 4 floats at a time (baseline for x86_64)
-/// - **Scalar**: Fallback for other platforms
-///
-/// This dedicated dot product function is more efficient than
-/// [`cosine_similarity`] when you only need the dot product and not the
-/// magnitudes (e.g., when working with pre-normalized vectors).
-#[inline]
-pub fn dot_product(a: &[f32], b: &[f32]) -> Result<f32> {
-    check_dimensions_match(a, b)?;
-
-    // Handle empty vectors
-    if a.is_empty() {
-        return Ok(0.0);
-    }
-
-    // Use SIMD-accelerated computation when available
-    Ok(dot_product_sum(a, b))
-}
-
-// ============================================================================
-// Normalization
-// ============================================================================
-
-/// Computes the magnitude (L2 norm) of a vector.
-///
-/// The magnitude is the square root of the sum of squared elements:
-/// `||v|| = sqrt(v₀² + v₁² + ... + vₙ²)`
-///
-/// # Arguments
-///
-/// * `v` - The vector to compute the magnitude of
-///
-/// # Returns
-///
-/// The magnitude as a non-negative f32. Returns 0.0 for empty vectors.
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::magnitude;
-///
-/// // Classic 3-4-5 right triangle
-/// let v = vec![3.0, 4.0];
-/// let mag = magnitude(&v);
-/// assert!((mag - 5.0).abs() < 1e-6);
-///
-/// // Unit vector has magnitude 1
-/// let unit = vec![1.0, 0.0, 0.0];
-/// assert!((magnitude(&unit) - 1.0).abs() < 1e-6);
-/// ```
-///
-/// # Performance
-///
-/// This function uses SIMD-accelerated dot product internally:
-/// `magnitude(v) = sqrt(dot_product(v, v))`
-#[inline(always)]
-pub fn magnitude(v: &[f32]) -> f32 {
-    if v.is_empty() {
-        return 0.0;
-    }
-    dot_product_sum(v, v).sqrt()
-}
-
-/// Computes the squared magnitude of a vector.
-///
-/// This is equivalent to `magnitude(v).powi(2)` but avoids the square root,
-/// making it faster for comparisons where the actual magnitude isn't needed.
-///
-/// # Arguments
-///
-/// * `v` - The vector to compute the squared magnitude of
-///
-/// # Returns
-///
-/// The squared magnitude as a non-negative f32. Returns 0.0 for empty vectors.
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::squared_magnitude;
-///
-/// let v = vec![3.0, 4.0];
-/// let sq_mag = squared_magnitude(&v);
-/// assert!((sq_mag - 25.0).abs() < 1e-6); // 3² + 4² = 25
-/// ```
-#[inline(always)]
-pub fn squared_magnitude(v: &[f32]) -> f32 {
-    if v.is_empty() {
-        return 0.0;
-    }
-    dot_product_sum(v, v)
-}
-
-/// Normalizes a vector to unit length (L2 normalization).
-///
-/// Creates a new vector with the same direction but magnitude 1.0.
-/// For zero vectors (magnitude = 0), returns a zero vector of the same length.
-///
-/// # Arguments
-///
-/// * `v` - The vector to normalize
-///
-/// # Returns
-///
-/// A new `Vec<f32>` with unit length, or a zero vector if the input is zero.
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::{normalize, magnitude};
-///
-/// let v = vec![3.0, 4.0];
-/// let unit = normalize(&v);
-///
-/// // Normalized vector has magnitude 1
-/// assert!((magnitude(&unit) - 1.0).abs() < 1e-6);
-///
-/// // Direction is preserved: [3, 4] -> [0.6, 0.8]
-/// assert!((unit[0] - 0.6).abs() < 1e-6);
-/// assert!((unit[1] - 0.8).abs() < 1e-6);
-///
-/// // Zero vector stays zero
-/// let zero = vec![0.0, 0.0];
-/// let normalized_zero = normalize(&zero);
-/// assert_eq!(normalized_zero, vec![0.0, 0.0]);
-/// ```
-///
-/// # Performance
-///
-/// This function allocates a new vector. For in-place normalization without
-/// allocation, use [`normalize_in_place`].
-/// Uses SIMD-accelerated scalar multiplication (AVX2/SSE2) for optimal performance.
-///
-/// # Note on Dimension Validation
-///
-/// Unlike two-vector functions like [`cosine_similarity`], normalization functions
-/// do not validate against `MAX_VECTOR_DIMENSIONS`. This is intentional because:
-/// - Single-vector operations don't have dimension mismatch issues
-/// - Dimension limits are enforced at storage time (see [`PropertyValue::vector`])
-/// - Additional checks would add overhead without safety benefit
-#[inline]
-pub fn normalize(v: &[f32]) -> Vec<f32> {
-    let sq_mag = squared_magnitude(v);
-    // Use squared magnitude threshold to avoid denormal number issues.
-    // See SQUARED_MAGNITUDE_THRESHOLD for details.
-    if sq_mag < SQUARED_MAGNITUDE_THRESHOLD {
-        // Return zero vector of same length
-        return vec![0.0; v.len()];
-    }
-    // Copy then scale in place using SIMD
-    // Compute 1/sqrt(sq_mag) directly to avoid intermediate variable
-    let mut result: Vec<f32> = v.to_vec();
-    let inv_mag = 1.0 / sq_mag.sqrt();
-    scale_in_place(&mut result, inv_mag);
-    result
-}
-
-/// Normalizes a vector to unit length in place.
-///
-/// Modifies the vector in place to have magnitude 1.0.
-/// For zero vectors (magnitude = 0), leaves the vector unchanged.
-///
-/// # Arguments
-///
-/// * `v` - The vector to normalize (modified in place)
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::{normalize_in_place, magnitude};
-///
-/// let mut v = vec![3.0, 4.0];
-/// normalize_in_place(&mut v);
-///
-/// // Now has magnitude 1
-/// assert!((magnitude(&v) - 1.0).abs() < 1e-6);
-///
-/// // Direction is preserved: [3, 4] -> [0.6, 0.8]
-/// assert!((v[0] - 0.6).abs() < 1e-6);
-/// assert!((v[1] - 0.8).abs() < 1e-6);
-/// ```
-///
-/// # Performance
-///
-/// This function modifies the vector in place without allocation, making it
-/// more efficient than [`normalize`] when a new vector isn't needed.
-/// Uses SIMD-accelerated scalar multiplication (AVX2/SSE2) for optimal performance.
-#[inline]
-pub fn normalize_in_place(v: &mut [f32]) {
-    let sq_mag = squared_magnitude(v);
-    // Use squared magnitude threshold to avoid denormal number issues.
-    // See SQUARED_MAGNITUDE_THRESHOLD for details.
-    if sq_mag < SQUARED_MAGNITUDE_THRESHOLD {
-        // Leave zero/near-zero vector unchanged
-        return;
-    }
-    // Compute 1/sqrt(sq_mag) directly to avoid intermediate variable
-    let inv_mag = 1.0 / sq_mag.sqrt();
-    scale_in_place(v, inv_mag);
-}
-
-/// Checks if a vector is normalized (has magnitude approximately 1.0).
-///
-/// This is useful for validating that vectors are properly normalized before
-/// using optimized functions like [`cosine_similarity_normalized`].
-///
-/// # Arguments
-///
-/// * `v` - The vector to check
-/// * `tolerance` - Maximum allowed deviation from 1.0 (e.g., 1e-6)
-///
-/// # Returns
-///
-/// `true` if the magnitude is within `tolerance` of 1.0, `false` otherwise.
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::{is_normalized, normalize};
-///
-/// let v = vec![3.0, 4.0];
-/// assert!(!is_normalized(&v, 1e-6));
-///
-/// let unit = normalize(&v);
-/// assert!(is_normalized(&unit, 1e-6));
-/// ```
-#[inline]
-pub fn is_normalized(v: &[f32], tolerance: f32) -> bool {
-    debug_assert!(
-        (0.0..1.0).contains(&tolerance),
-        "tolerance must be in range [0.0, 1.0), got {}",
-        tolerance
-    );
-    // Use squared_magnitude to avoid sqrt for better numerical stability
-    // |magnitude - 1.0| <= tolerance  ⟺  (1-tolerance)² <= ||v||² <= (1+tolerance)²
-    let sq_mag = squared_magnitude(v);
-    let lower = (1.0 - tolerance).max(0.0).powi(2);
-    let upper = (1.0 + tolerance).powi(2);
-    sq_mag >= lower && sq_mag <= upper
-}
-
-/// Checks if a vector is normalized using the default tolerance.
-///
-/// This is a convenience wrapper around [`is_normalized`] that uses
-/// [`NORMALIZATION_TOLERANCE`] (1e-6) as the tolerance value.
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::{is_normalized_default, normalize};
-///
-/// let v = vec![3.0, 4.0];
-/// assert!(!is_normalized_default(&v));
-///
-/// let unit = normalize(&v);
-/// assert!(is_normalized_default(&unit));
-/// ```
-#[inline]
-pub fn is_normalized_default(v: &[f32]) -> bool {
-    is_normalized(v, NORMALIZATION_TOLERANCE)
-}
-
-// ============================================================================
-// Vector Validation
-// ============================================================================
-
-/// Validates that a vector contains no NaN or Infinity values.
-///
-/// This function scans all elements of the vector and returns an error if any
-/// invalid floating-point values are found. NaN and Infinity values would cause
-/// incorrect results in distance/similarity calculations and should be caught early.
-///
-/// # Arguments
-///
-/// * `v` - The vector slice to validate
-///
-/// # Returns
-///
-/// * `Ok(())` if all elements are valid finite numbers
-/// * `Err(VectorError::ContainsNaN)` if any NaN values are found
-/// * `Err(VectorError::ContainsInfinity)` if any Infinity values are found (checked after NaN)
-///
-/// # Note
-///
-/// NaN is checked first, so if a vector contains both NaN and Infinity values,
-/// the NaN error will be returned. This is because NaN values are generally more
-/// problematic (NaN != NaN, propagates through calculations).
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::validate_vector;
-///
-/// // Valid vector
-/// let v = vec![1.0, 2.0, 3.0];
-/// assert!(validate_vector(&v).is_ok());
-///
-/// // Vector with NaN
-/// let v_nan = vec![1.0, f32::NAN, 3.0];
-/// assert!(validate_vector(&v_nan).is_err());
-///
-/// // Vector with Infinity
-/// let v_inf = vec![1.0, f32::INFINITY, 3.0];
-/// assert!(validate_vector(&v_inf).is_err());
-///
-/// // Empty vector is valid
-/// let empty: Vec<f32> = vec![];
-/// assert!(validate_vector(&empty).is_ok());
-/// ```
-#[inline]
-pub fn validate_vector(v: &[f32]) -> Result<()> {
-    // Use a single pass to count both NaN and Infinity values for efficiency.
-    // This is more efficient than iterating twice, especially for large vectors.
-    let (nan_count, inf_count) = v.iter().fold((0usize, 0usize), |(nan, inf), &val| {
-        if val.is_nan() {
-            (nan + 1, inf)
-        } else if val.is_infinite() {
-            (nan, inf + 1)
-        } else {
-            (nan, inf)
-        }
-    });
-
-    // Per the function's contract, NaN is checked first.
-    if nan_count > 0 {
-        return Err(Error::Vector(VectorError::ContainsNaN { count: nan_count }));
-    }
-
-    if inf_count > 0 {
-        return Err(Error::Vector(VectorError::ContainsInfinity {
-            count: inf_count,
-        }));
-    }
-
-    Ok(())
-}
-
-/// Checks that two vectors have matching dimensions.
-///
-/// Many vector operations require vectors of equal length. This function provides
-/// a convenient way to validate dimension compatibility before performing operations.
-///
-/// # Arguments
-///
-/// * `a` - The first vector (its length is considered the "expected" dimension)
-/// * `b` - The second vector (its length is compared against `a`)
-///
-/// # Returns
-///
-/// * `Ok(())` if both vectors have the same length
-/// * `Err(VectorError::DimensionMismatch)` if lengths differ
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::check_dimensions_match;
-///
-/// let v1 = vec![1.0, 2.0, 3.0];
-/// let v2 = vec![4.0, 5.0, 6.0];
-/// let v3 = vec![1.0, 2.0];
-///
-/// // Same dimensions - OK
-/// assert!(check_dimensions_match(&v1, &v2).is_ok());
-///
-/// // Different dimensions - Error
-/// assert!(check_dimensions_match(&v1, &v3).is_err());
-///
-/// // Empty vectors match
-/// let empty1: Vec<f32> = vec![];
-/// let empty2: Vec<f32> = vec![];
-/// assert!(check_dimensions_match(&empty1, &empty2).is_ok());
-/// ```
-#[inline]
-pub fn check_dimensions_match(a: &[f32], b: &[f32]) -> Result<()> {
-    if a.len() != b.len() {
-        return Err(Error::Vector(VectorError::DimensionMismatch {
-            expected: a.len(),
-            actual: b.len(),
-        }));
-    }
-    Ok(())
-}
-
-/// Validates a vector and checks that its dimension is within bounds.
-///
-/// This is a convenience function that combines validation (NaN/Infinity checking)
-/// with dimension bounds checking. Useful when processing user-provided vectors
-/// that need both validation and size constraints.
-///
-/// # Arguments
-///
-/// * `v` - The vector slice to validate
-/// * `max_dimension` - The maximum allowed dimension (length)
-///
-/// # Returns
-///
-/// * `Ok(())` if the vector is valid and within dimension bounds
-/// * `Err(VectorError::ContainsNaN)` if any NaN values are found
-/// * `Err(VectorError::ContainsInfinity)` if any Infinity values are found
-/// * `Err(VectorError::DimensionTooLarge)` if the vector exceeds max_dimension
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::validate_vector_with_bounds;
-///
-/// // Valid vector within bounds
-/// let v = vec![1.0, 2.0, 3.0];
-/// assert!(validate_vector_with_bounds(&v, 10).is_ok());
-///
-/// // Vector too large
-/// assert!(validate_vector_with_bounds(&v, 2).is_err());
-///
-/// // Vector with invalid values
-/// let v_nan = vec![1.0, f32::NAN];
-/// assert!(validate_vector_with_bounds(&v_nan, 10).is_err());
-/// ```
-#[inline]
-pub fn validate_vector_with_bounds(v: &[f32], max_dimension: usize) -> Result<()> {
-    // Check dimension first (fast check)
-    if v.len() > max_dimension {
-        return Err(Error::Vector(VectorError::DimensionTooLarge {
-            dimension: v.len(),
-            max_allowed: max_dimension,
-        }));
-    }
-
-    // Then validate contents
-    validate_vector(v)
-}
-
-// ============================================================================
-// Sparse Vector Similarity Functions
-// ============================================================================
-
-/// Computes the dot product between two sparse vectors.
-///
-/// The dot product is computed efficiently by iterating only over non-zero
-/// elements. Vectors with different dimensions are handled gracefully by
-/// only considering indices that exist in both vectors.
-///
-/// # Arguments
-///
-/// * `a` - First sparse vector
-/// * `b` - Second sparse vector
-///
-/// # Returns
-///
-/// * `Ok(f32)` - The dot product
-/// * `Err` - Never returns an error (kept for API consistency with dense vectors)
-///
-/// # Algorithm
-///
-/// Uses a merge-like algorithm since indices are sorted:
-/// 1. Maintain pointers to both index arrays
-/// 2. When indices match, multiply values and add to sum
-/// 3. Advance pointer with smaller index
-/// 4. Time complexity: O(nnz_a + nnz_b)
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::{SparseVec, sparse_dot_product};
-///
-/// // Sparse vectors: [1, 0, 2, 0, 0] and [0, 0, 3, 0, 4]
-/// let a = SparseVec::new(vec![0, 2], vec![1.0, 2.0], 5).unwrap();
-/// let b = SparseVec::new(vec![2, 4], vec![3.0, 4.0], 5).unwrap();
-///
-/// // Only index 2 overlaps: 2.0 * 3.0 = 6.0
-/// let dot = sparse_dot_product(&a, &b).unwrap();
-/// assert_eq!(dot, 6.0);
-/// ```
-///
-/// # Performance
-///
-/// For vectors with nnz_a and nnz_b non-zero elements:
-/// - Time: O(nnz_a + nnz_b) - linear in sparsity
-/// - Space: O(1) - no allocations
-/// - Much faster than dense dot product when vectors are sparse
-pub fn sparse_dot_product(a: &SparseVec, b: &SparseVec) -> Result<f32> {
-    let mut sum = 0.0f32;
-    let mut i = 0;
-    let mut j = 0;
-
-    let a_indices = a.indices();
-    let a_values = a.values();
-    let b_indices = b.indices();
-    let b_values = b.values();
-
-    // Merge-like algorithm since indices are sorted
-    while i < a_indices.len() && j < b_indices.len() {
-        if a_indices[i] == b_indices[j] {
-            // Indices match - multiply and add
-            sum += a_values[i] * b_values[j];
-            i += 1;
-            j += 1;
-        } else if a_indices[i] < b_indices[j] {
-            // a's index is smaller, advance a
-            i += 1;
-        } else {
-            // b's index is smaller, advance b
-            j += 1;
-        }
-    }
-
-    Ok(sum)
-}
-
-/// Computes cosine similarity between two sparse vectors.
-///
-/// Cosine similarity measures the angle between vectors, ranging from -1 (opposite)
-/// to 1 (identical direction). For sparse vectors, this is computed efficiently
-/// by only considering non-zero elements.
-///
-/// # Arguments
-///
-/// * `a` - First sparse vector
-/// * `b` - Second sparse vector
-///
-/// # Returns
-///
-/// * `Ok(f32)` - Cosine similarity in range [-1, 1]
-/// * `Err` - If either vector has zero magnitude
-///
-/// # Formula
-///
-/// ```text
-/// cosine_similarity = dot(a, b) / (||a|| * ||b||)
-/// ```
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::{SparseVec, sparse_cosine_similarity};
-///
-/// let a = SparseVec::new(vec![0, 2], vec![1.0, 1.0], 5).unwrap();
-/// let b = SparseVec::new(vec![0, 2], vec![1.0, 1.0], 5).unwrap();
-///
-/// // Identical vectors have cosine similarity = 1.0
-/// let sim = sparse_cosine_similarity(&a, &b).unwrap();
-/// assert!((sim - 1.0).abs() < 1e-6);
-/// ```
-///
-/// # Performance
-///
-/// - Time: O(nnz_a + nnz_b) for dot product + O(nnz_a + nnz_b) for magnitudes
-/// - Space: O(1)
-/// - Much faster than dense cosine for sparse data
-pub fn sparse_cosine_similarity(a: &SparseVec, b: &SparseVec) -> Result<f32> {
-    let dot = sparse_dot_product(a, b)?;
-    let mag_a = a.magnitude();
-    let mag_b = b.magnitude();
-
-    // Handle zero magnitude vectors
-    if mag_a < SQUARED_MAGNITUDE_THRESHOLD || mag_b < SQUARED_MAGNITUDE_THRESHOLD {
-        return Ok(0.0);
-    }
-
-    let similarity = dot / (mag_a * mag_b);
-
-    // Clamp to [-1, 1] to handle floating-point errors
-    Ok(similarity.clamp(-1.0, 1.0))
-}
-
-/// Computes squared Euclidean distance between two sparse vectors.
-///
-/// The squared distance is more efficient than computing the full Euclidean
-/// distance since it avoids the square root operation. For comparing distances,
-/// squared distance preserves ordering.
-///
-/// # Arguments
-///
-/// * `a` - First sparse vector
-/// * `b` - Second sparse vector
-///
-/// # Returns
-///
-/// * `Ok(f32)` - Squared Euclidean distance
-/// * `Err(VectorError::DimensionMismatch)` - If vectors have different dimensions
-///
-/// # Formula
-///
-/// ```text
-/// squared_distance = ||a - b||² = Σ(a_i - b_i)²
-/// ```
-///
-/// For sparse vectors, we only need to compute:
-/// - ||a||² + ||b||² - 2*dot(a,b)
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::{SparseVec, sparse_squared_euclidean_distance};
-///
-/// let a = SparseVec::new(vec![0], vec![3.0], 5).unwrap();
-/// let b = SparseVec::new(vec![], vec![], 5).unwrap(); // Zero vector
-///
-/// // Distance from [3,0,0,0,0] to [0,0,0,0,0] = 9
-/// let dist_sq = sparse_squared_euclidean_distance(&a, &b).unwrap();
-/// assert_eq!(dist_sq, 9.0);
-/// ```
-///
-/// # Performance
-///
-/// - Time: O(nnz_a + nnz_b)
-/// - Space: O(1)
-pub fn sparse_squared_euclidean_distance(a: &SparseVec, b: &SparseVec) -> Result<f32> {
-    // Check dimensions match
-    if a.dimension() != b.dimension() {
-        return Err(VectorError::DimensionMismatch {
-            expected: a.dimension(),
-            actual: b.dimension(),
-        }
-        .into());
-    }
-
-    // ||a - b||² = ||a||² + ||b||² - 2*dot(a,b)
-    let dot = sparse_dot_product(a, b)?;
-    let mag_a_sq = a.squared_magnitude();
-    let mag_b_sq = b.squared_magnitude();
-
-    Ok(mag_a_sq + mag_b_sq - 2.0 * dot)
-}
-
-/// Computes Euclidean distance between two sparse vectors.
-///
-/// This is the L2 distance - the straight-line distance between two points.
-///
-/// # Arguments
-///
-/// * `a` - First sparse vector
-/// * `b` - Second sparse vector
-///
-/// # Returns
-///
-/// * `Ok(f32)` - Euclidean distance
-/// * `Err(VectorError::DimensionMismatch)` - If vectors have different dimensions
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::{SparseVec, sparse_euclidean_distance};
-///
-/// let a = SparseVec::new(vec![0], vec![3.0], 5).unwrap();
-/// let b = SparseVec::new(vec![], vec![], 5).unwrap(); // Zero vector
-///
-/// // Distance from [3,0,0,0,0] to [0,0,0,0,0] = 3.0
-/// let dist = sparse_euclidean_distance(&a, &b).unwrap();
-/// assert_eq!(dist, 3.0);
-/// ```
-#[inline]
-pub fn sparse_euclidean_distance(a: &SparseVec, b: &SparseVec) -> Result<f32> {
-    sparse_squared_euclidean_distance(a, b).map(|sq| sq.sqrt())
-}
-
-// ============================================================================
-// Distance Metric Enum
-// ============================================================================
-
-/// Specifies which distance or similarity metric to use for vector operations.
-///
-/// This enum provides a unified interface for computing distances and similarities
-/// between vectors, dispatching to the appropriate underlying function based on the
-/// selected metric.
-///
-/// # Choosing a Metric
-///
-/// | Metric | Best For | Range | Notes |
-/// |--------|----------|-------|-------|
-/// | [`Cosine`](DistanceMetric::Cosine) | Semantic similarity, text embeddings | [-1, 1] similarity, [0, 2] distance | Scale-invariant, most common for embeddings |
-/// | [`Euclidean`](DistanceMetric::Euclidean) | Spatial data, image features | [0, ∞) distance | Sensitive to vector magnitude |
-/// | [`DotProduct`](DistanceMetric::DotProduct) | Pre-normalized vectors, MaxIP search | (-∞, ∞) | Fastest; requires normalized vectors for cosine-like behavior |
-///
-/// # Example
-///
-/// ```rust
-/// use gallifreydb::core::vector::DistanceMetric;
-///
-/// let a = vec![1.0, 0.0, 0.0];
-/// let b = vec![0.0, 1.0, 0.0];
-///
-/// // Using cosine similarity (orthogonal vectors = 0 similarity)
-/// let similarity = DistanceMetric::Cosine.compute_similarity(&a, &b).unwrap();
-/// assert!((similarity - 0.0).abs() < 1e-6);
-///
-/// // Using euclidean distance
-/// let distance = DistanceMetric::Euclidean.compute_distance(&a, &b).unwrap();
-/// assert!((distance - std::f32::consts::SQRT_2).abs() < 1e-6);
-/// ```
-///
-/// # Performance
-///
-/// All metrics use SIMD acceleration (AVX2/SSE2) when available. For maximum
-/// performance with large-scale similarity search:
-///
-/// 1. Pre-normalize vectors with [`normalize`] or [`normalize_in_place`]
-/// 2. Use [`DotProduct`](DistanceMetric::DotProduct) metric (single SIMD operation)
-/// 3. Store normalized vectors to avoid repeated normalization
-///
-/// # Future Enhancements
-///
-/// - **Serialization**: When serde is added as a dependency, this enum will
-///   support `#[serde(rename_all = "snake_case")]` for JSON/config serialization
-/// - **Batch operations**: `compute_distances_batch()` for SIMD-optimized
-///   multi-vector distance computation
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
-pub enum DistanceMetric {
-    /// Cosine similarity/distance.
-    ///
-    /// Measures the cosine of the angle between two vectors, making it
-    /// **scale-invariant** (only direction matters, not magnitude).
-    ///
-    /// - **Similarity**: Range [-1, 1] where 1 = identical direction,
-    ///   0 = orthogonal, -1 = opposite direction
-    /// - **Distance**: Computed as `1 - similarity`, range [0, 2]
-    ///
-    /// # When to Use
-    ///
-    /// - Text embeddings (word2vec, BERT, OpenAI embeddings)
-    /// - Semantic similarity where magnitude doesn't matter
-    /// - When vectors may have different scales
-    ///
-    /// # Implementation Note
-    ///
-    /// Uses [`cosine_similarity`] internally, which handles zero vectors
-    /// by returning 0.0 similarity.
-    #[default]
-    Cosine,
-
-    /// Euclidean (L2) distance.
-    ///
-    /// Measures the straight-line distance between two points in vector space.
-    /// Also known as the L2 norm of the difference vector.
-    ///
-    /// - **Distance**: Range [0, ∞) where 0 = identical vectors
-    /// - **Similarity**: Computed as `1 / (1 + distance)`, range (0, 1]
-    ///
-    /// # When to Use
-    ///
-    /// - Spatial data (coordinates, positions)
-    /// - Image feature vectors
-    /// - When absolute magnitude differences matter
-    /// - K-means clustering (uses squared Euclidean internally)
-    ///
-    /// # Implementation Note
-    ///
-    /// Uses [`euclidean_distance`] internally, which uses SIMD-accelerated
-    /// squared difference computation.
-    Euclidean,
-
-    /// Inner (dot) product.
-    ///
-    /// Computes the sum of element-wise products. For normalized vectors,
-    /// this equals cosine similarity but is faster (single SIMD operation).
-    ///
-    /// - **Raw value**: Range (-∞, ∞)
-    /// - **Similarity**: Raw dot product value (higher = more similar)
-    /// - **Distance**: Computed as `1 - dot_product`, which is meaningful
-    ///   only for normalized vectors
-    ///
-    /// # When to Use
-    ///
-    /// - **Maximum Inner Product Search (MIPS)**: When you want the highest
-    ///   dot product, not necessarily the closest vector
-    /// - **Pre-normalized vectors**: Equivalent to cosine but faster
-    /// - **Learned embeddings**: Some models are trained with dot product loss
-    ///
-    /// # Important
-    ///
-    /// For non-normalized vectors, dot product is **not** a proper distance
-    /// metric (doesn't satisfy triangle inequality). Use [`Cosine`](DistanceMetric::Cosine)
-    /// for general similarity or ensure vectors are normalized first.
-    DotProduct,
-}
-
-impl DistanceMetric {
-    /// Computes the distance between two vectors using this metric.
-    ///
-    /// Lower values indicate more similar vectors.
-    ///
-    /// # Returns
-    ///
-    /// - [`Cosine`](DistanceMetric::Cosine): `1 - cosine_similarity`, range [0, 2]
-    /// - [`Euclidean`](DistanceMetric::Euclidean): L2 distance, range [0, ∞)
-    /// - [`DotProduct`](DistanceMetric::DotProduct): `1 - dot_product` (meaningful only for normalized vectors)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the vectors have different lengths.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::DistanceMetric;
-    ///
-    /// let a = vec![1.0, 0.0];
-    /// let b = vec![1.0, 0.0];
-    ///
-    /// // Identical vectors have zero distance
-    /// assert!((DistanceMetric::Cosine.compute_distance(&a, &b).unwrap() - 0.0).abs() < 1e-6);
-    /// assert!((DistanceMetric::Euclidean.compute_distance(&a, &b).unwrap() - 0.0).abs() < 1e-6);
-    /// ```
-    #[inline]
-    pub fn compute_distance(&self, a: &[f32], b: &[f32]) -> Result<f32> {
-        match self {
-            DistanceMetric::Cosine => cosine_similarity(a, b).map(|sim| 1.0 - sim),
-            DistanceMetric::Euclidean => euclidean_distance(a, b),
-            DistanceMetric::DotProduct => dot_product(a, b).map(|dp| 1.0 - dp),
-        }
-    }
-
-    /// Computes the similarity between two vectors using this metric.
-    ///
-    /// Higher values indicate more similar vectors.
-    ///
-    /// # Returns
-    ///
-    /// - [`Cosine`](DistanceMetric::Cosine): Cosine similarity, range [-1, 1]
-    /// - [`Euclidean`](DistanceMetric::Euclidean): `1 / (1 + distance)`, range (0, 1]
-    /// - [`DotProduct`](DistanceMetric::DotProduct): Raw dot product, range (-∞, ∞)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the vectors have different lengths.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::DistanceMetric;
-    ///
-    /// let a = vec![1.0, 0.0];
-    /// let b = vec![1.0, 0.0];
-    ///
-    /// // Identical vectors have maximum similarity
-    /// assert!((DistanceMetric::Cosine.compute_similarity(&a, &b).unwrap() - 1.0).abs() < 1e-6);
-    /// assert!((DistanceMetric::Euclidean.compute_similarity(&a, &b).unwrap() - 1.0).abs() < 1e-6);
-    /// ```
-    #[inline]
-    pub fn compute_similarity(&self, a: &[f32], b: &[f32]) -> Result<f32> {
-        match self {
-            DistanceMetric::Cosine => cosine_similarity(a, b),
-            DistanceMetric::Euclidean => euclidean_distance(a, b).map(|dist| 1.0 / (1.0 + dist)),
-            DistanceMetric::DotProduct => dot_product(a, b),
-        }
-    }
-
-    /// Returns a human-readable name for this metric.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::DistanceMetric;
-    ///
-    /// assert_eq!(DistanceMetric::Cosine.name(), "cosine");
-    /// assert_eq!(DistanceMetric::Euclidean.name(), "euclidean");
-    /// assert_eq!(DistanceMetric::DotProduct.name(), "dot_product");
-    /// ```
-    #[inline]
-    pub const fn name(&self) -> &'static str {
-        match self {
-            DistanceMetric::Cosine => "cosine",
-            DistanceMetric::Euclidean => "euclidean",
-            DistanceMetric::DotProduct => "dot_product",
-        }
-    }
-
-    /// Returns whether this metric requires normalized vectors for optimal results.
-    ///
-    /// - [`Cosine`](DistanceMetric::Cosine): No (handles normalization internally)
-    /// - [`Euclidean`](DistanceMetric::Euclidean): No (works with any vectors)
-    /// - [`DotProduct`](DistanceMetric::DotProduct): Yes (otherwise not a proper similarity)
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use gallifreydb::core::vector::DistanceMetric;
-    ///
-    /// assert!(!DistanceMetric::Cosine.requires_normalized_vectors());
-    /// assert!(!DistanceMetric::Euclidean.requires_normalized_vectors());
-    /// assert!(DistanceMetric::DotProduct.requires_normalized_vectors());
-    /// ```
-    #[inline]
-    pub const fn requires_normalized_vectors(&self) -> bool {
-        matches!(self, DistanceMetric::DotProduct)
-    }
-}
-
-impl fmt::Display for DistanceMetric {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.name())
-    }
-}
-
-// ============================================================================
-// Tests
-// ============================================================================
-
-#[cfg(test)]
-mod tests {
+use crate::utils::error::{Error, VectorError};
+mod unit_tests {
     use super::*;
 
     #[test]
@@ -4344,22 +1825,22 @@ mod tests {
         if is_x86_feature_detected!("sse2") {
             unsafe {
                 // dot_and_magnitudes_sse2
-                let (dot, mag_a, mag_b) = super::simd::dot_and_magnitudes_sse2(&a, &b);
+                let (dot, mag_a, mag_b) = super::simd::internal::dot_and_magnitudes_sse2(&a, &b);
                 assert!(!dot.is_nan());
                 assert!(mag_a > 0.0);
                 assert!(mag_b > 0.0);
 
                 // dot_product_sse2
-                let dot_prod = super::simd::dot_product_sse2(&a, &b);
+                let dot_prod = super::simd::internal::dot_product_sse2(&a, &b);
                 assert!((dot - dot_prod).abs() < 1e-5);
 
                 // squared_diff_sum_sse2
-                let sq_diff = super::simd::squared_diff_sum_sse2(&a, &b);
+                let sq_diff = super::simd::internal::squared_diff_sum_sse2(&a, &b);
                 assert!(sq_diff >= 0.0);
 
                 // scale_in_place_sse2
                 let mut v = a.clone();
-                super::simd::scale_in_place_sse2(&mut v, 2.0);
+                super::simd::internal::scale_in_place_sse2(&mut v, 2.0);
                 assert!((v[0] - 2.0).abs() < 1e-5);
             }
         }
@@ -4369,22 +1850,22 @@ mod tests {
             unsafe {
                 // scale_in_place_avx2 only needs AVX2
                 let mut v = a.clone();
-                super::simd::scale_in_place_avx2(&mut v, 2.0);
+                super::simd::internal::scale_in_place_avx2(&mut v, 2.0);
                 assert!((v[0] - 2.0).abs() < 1e-5);
 
                 // Others need FMA + AVX2
                 if is_x86_feature_detected!("fma") {
                     // dot_and_magnitudes_avx2
-                    let (dot, mag_a, _mag_b) = super::simd::dot_and_magnitudes_avx2(&a, &b);
+                    let (dot, mag_a, _mag_b) = super::simd::internal::dot_and_magnitudes_avx2(&a, &b);
                     assert!(!dot.is_nan());
                     assert!(mag_a > 0.0);
 
                     // dot_product_avx2
-                    let dot_prod = super::simd::dot_product_avx2(&a, &b);
+                    let dot_prod = super::simd::internal::dot_product_avx2(&a, &b);
                     assert!((dot - dot_prod).abs() < 1e-5);
 
                     // squared_diff_sum_avx2
-                    let sq_diff = super::simd::squared_diff_sum_avx2(&a, &b);
+                    let sq_diff = super::simd::internal::squared_diff_sum_avx2(&a, &b);
                     assert!(sq_diff >= 0.0);
                 }
             }
